### 2020-01-01 #356 @pat-s
Submission: circle
Submitting Author: Patrick Schratz (@pat-s)
Repository: https://github.com/ropenscilabs/circle
Version submitted: v0.5.0.9001
Version accepted: TBD

---

```
Package: circle
Title: R Client Package for Circle CI
Version: 0.7.0
    person(given = "Patrick",
           family = "Schratz",
           email = "patrick.schratz@gmail.com",
Description: Tools for interacting with the 'Circle CI' API.
    Besides executing common tasks such as querying build logs and
    restarting builds, this package also helps setting up permissions to
    deploy from builds.
    https://github.com/ropenscilabs/circle
    cli (>= 2.0.0),
    httr,
    jsonlite
    clipr,
    covr,
    gert,
    gh,
    openssl,
    usethis (>= 2.0.0),
    vcr,
    withr
    knitr
Config/testthat/parallel: false

```

It helps to automate CI workflows and goes in line with the {tic} and {travis} package.

Users who prefer to interact with the command line to query information about CI build statuses.
No scientific applications.

Besides a [discontinued package](https://github.com/cloudyr/circleci) which is 4 years old, there is no similar package.
