### 2020-01-13 221bc17
export set_env_var()
