+++ NAMESPACE
@@ -19,6 +19,7 @@ export(github_repo)
 export(has_checkout_key)
 export(list_projects)
 export(new_build)
+export(set_env_var)
 export(use_circle_deploy)
 import(cli)
 import(httr)
