+++ DESCRIPTION
@@ -1,6 +1,6 @@
 Package: circle
 Title: R client package for the Circle CI API
-Version: 0.5.0.9000
+Version: 0.5.0.9001
 Authors@R: 
     person(given = "Patrick",
            family = "Schratz",
