### 2020-01-17 90576b2
stop_for_status when checkout key cannot be created
