### 2020-01-17 935204c
check for user-key instead of github-user-key
