+++ DESCRIPTION
@@ -30,7 +30,8 @@ Suggests:
     rmarkdown,
     testthat,
     usethis,
-    utils
+    utils,
+    withr
 VignetteBuilder: 
     knitr
 Encoding: UTF-8
