### 2020-01-22 0d401df
tests: skip_on_cran(), use with_dir() approach
