+++ tests/testthat/test-print.R
@@ -1,7 +1,9 @@
 context("print S3 methods")
 
 test_that("s3 print method for 'circle_pipeline' works", {
-  out <- get_pipelines(repo = repo, user = user)
+  skip_on_cran()
+
+  out <- get_pipelines()
   capture.output(expect_message(
     print(out),
     "A collection of 20 Circle CI pipelines"
@@ -9,7 +11,10 @@ test_that("s3 print method for 'circle_pipeline' works", {
 })
 
 test_that("s3 print method for 'circle_collection' works if class != 'circle_pipelines'", { # nolint
-  out <- get_workflows(repo = repo, user = user)
+
+  skip_on_cran()
+
+  out <- get_workflows()
   capture.output(expect_message(
     print(out),
     "A collection of 10 Circle CI workflows"
@@ -17,7 +22,9 @@ test_that("s3 print method for 'circle_collection' works if class != 'circle_pip
 })
 
 test_that("s3 print method for 'circle_workflow' works", {
-  wf <- get_workflows(repo = repo, user = user)
+  skip_on_cran()
+
+  wf <- get_workflows()
   capture.output(expect_message(
     print(wf[[1]]),
     "A Circle CI workflow:"
@@ -25,7 +32,9 @@ test_that("s3 print method for 'circle_workflow' works", {
 })
 
 test_that("s3 print method for 'circle_job' works", {
-  job <- get_jobs(repo = repo, user = user)
+  skip_on_cran()
+
+  job <- get_jobs()
   capture.output(expect_message(
     print(job[[1]]),
     "A Circle CI job:"
