+++ tests/testthat/test-env-vars.R
@@ -1,30 +1,33 @@
-context("env-vars")
+withr::with_dir(
+  "travis-testthat",
+  {
+    test_that("setting env vars works", {
+      skip_on_cran()
 
-setwd("./travis-testthat")
+      expect_silent(
+        set_env_var(
+          list(foo = "test"),
+          quiet = TRUE
+        )
+      )
+    })
 
-test_that("setting env vars works", {
-  expect_silent(
-    set_env_var(
-      list(foo = "test"),
-      quiet = TRUE,
-      repo = repo, user = user
-    )
-  )
-})
+    test_that("getting env vars works", {
+      skip_on_cran()
 
-test_that("getting env vars works", {
-  expect_silent(
-    get_env_vars(
-      repo = repo, user = user
-    )
-  )
-})
+      expect_silent(
+        get_env_vars()
+      )
+    })
 
-test_that("deleting env vars works", {
-  expect_silent(
-    delete_env_var(
-      var = "foo", quiet = TRUE,
-      repo = repo, user = user
-    )
-  )
-})
+    test_that("deleting env vars works", {
+      skip_on_cran()
+
+      expect_silent(
+        delete_env_var(
+          var = "foo", quiet = TRUE
+        )
+      )
+    })
+  }
+)
