### 2020-01-22 96f0125
README: Add info about testing the package locally
