### 2020-01-22 646bf33
README: add scope of package
