### 2020-01-22 4e5b053
vignette: Optimize authentication section, mention to install the Circle CI app
