### 2020-01-22 #356 @pat-s
Hi Noam,

thanks for taking this on!

> My initial attempt to test this package failed

Tests are running on a test repo which is cloned during testing.
Since the repo lives in my account, other users won't have permissions to start/cancel builds.
We will also have to skip all tests on CRAN since there is no way I can set the API keys there.
The API keys are stored on the CI providers securely, so tests will run there for everyone.
We could potentially move the test repo to `ropenscilabs` though this would also just give a handful people access to it if they have local keys set up.
I'd say having the tests running on the CI for everyone is the best solution here.
Let me know if I overlook something.

Also for some tests I need a GITHUB_TOKEN with extended permissions (to delete a private GH user key of mine automatically which will be created by `use_circle_deploy()`).

I've added this piece of information to `CONTRIBUTING.md` as suggested and put a link into the README.

> Also, to preempt another issue, your README is somewhat sparse and both it and the vignette lack an introduction to what the package is for and, importantly in this case, what it is not for.

I've added more information to the README.

In addition, I also optimized the Getting Started vignette. (You might want to wait a bit until the vignette is deployed because it seems there are some [SSL certificate issues](https://github.com/r-lib/pkgdown/issues/1157#issuecomment-577065966) with pkgdown today).

Let me know if there are still points to address before the review can start.
