+++ README.Rmd
@@ -28,7 +28,7 @@ This API version is still in beta and might undergo some changes in the near fut
 
 Some functions can also be used via API versions v1.1 and v1 via the `api_version` argument.
 However, this will only work if the respective endpoints are available for both API versions.
-There should be no need in practice to this.
+Usually, there should be no need in practice to fall back to API version < 2.
 
 If you want to get more information, have a look at the [document explaining changes between v1.1 and v2](https://github.com/CircleCI-Public/api-preview-docs/blob/master/docs/api-changes.md#endpoints-likely-being-removed-in-api-v2-still-available-in-v11-for-now).
 
