### 2020-01-22 63d20e5
sapply -> vapply
