+++ tests/testthat/test-checkout-keys.R
@@ -28,9 +28,13 @@ withr::with_dir(
       )
 
       # also delete GH SHH key
-      foo <- gh::gh("GET /user/keys", .token = Sys.getenv("GH_TOKEN_CHECKOUT_KEY"))
+      foo <- gh::gh("GET /user/keys",
+        .token = Sys.getenv("GH_TOKEN_CHECKOUT_KEY")
+      )
       # get id of last added key
-      key_id <- sapply(foo, function(x) x$id)[length(foo)]
+      key_id <- vapply(foo, function(x) x$id,
+        FUN.VALUE = integer(1)
+      )[length(foo)]
 
       gh::gh("DELETE /user/keys/:key_id",
         key_id = key_id,
