+++ vignettes/circle.Rmd
@@ -41,9 +41,9 @@ The easiest way to enable deployment from Circle CI builds to Github repositorie
 
 ```r
 circle::use_circle_deploy()
-✖ No 'github-user-key' found.
+✖ No 'user-key' found.
 ──────────────────────────────────────────────────────────────────────────────────────────────────
-✔ Added a 'github user key' to project '<repo-slug>' on Circle CI. 
+✔ Added a 'user key' to project '<repo-slug>' on Circle CI. 
 This enables deployment from builds.
 ```
 
