### 2020-01-24 #356 @noamross
Hi @pat-s, I don't think this is a great solution, because this means that a contributor or reviewer can't test the package, and if they do only pushing and having the PR run for you then you need to expose your extended-permission GITHUB_TOKEN (very insecure!) to all. I get that you will need to skip these on CRAN, but I'd much prefer that there be a full developer workflow. This should hopefully save you effort in the long run, as contributors will be able to test prior to submitting PRs.   It's fine if this takes some setting up by the contributor.

A solution could be that the user provides an env var `R_CIRCLE_TEST_REPO` (forked from your repo if its contents are important), as well as an extended permissions github token.
