### 2020-01-24 #356 @pat-s
> if they do only pushing and having the PR run for you then you need to expose your extended-permission GITHUB_TOKEN (very insecure!) to all

I don't see why anything should be exposed. The token is stored as a secure env var in the build and nothing will be exposed when a PR is run.

> This should hopefully save you effort in the long run, as contributors will be able to test prior to submitting PRs.

I see the advantage of being able to run the tests locally.
However, maybe your are overlooking the following: Even if an API key is set, the reviewers would need my API key to make API calls to `pat-s/travis-testthat` to start/cancel build.

Even when setting their personal API key, reviewers would need to

- adjust the test suite manually so tests run on a repo of their own (possibly create that first)
- repo needs to have XY builds because tests are not just running on the most recent build/job
- repo needs to have a working Circle CI workflow

Because testing an API client package is not trivial, I created the dummy repo to have a dedicated repo for testing API functionality.
Sometimes things go wrong and nobody wants to have such things happening on a production repo.

If I would be a reviewer of such a package, I would certainly dislike having to do all of this effort **just** to run tests locally whereas they just run fine on the CI.

I'd argue that a API client package is special when it comes to testing, especially when it comes to CI client packages for which it is not sufficient to just replace the API key.

From my perspective, it is important that the functions of this package work if users apply them to a repo of themselves (after having set the API key) rather than requiring them to adjust the whole test suite.

I am not sure if there was a previous case like this one - maybe a general decision how such packages are handled by {ropensci} is needed here.
The next one ({travis}) is already waiting.

Happy to make adjustments if I get a pointer how to do so.
I just do not see one right now.
