### 2020-01-27 #356 @noamross
A very large number of our packages are actually API clients; **circle**'s case is only a little different from several others. At a minimum we need a functioning test suite so that tests can pass locally and then later with rOpenSci's automatic checks.  But there are a number of ways to accomplish this.

- Split tests into those that run with and without certain tokens available, and selectively run tests based on availability.  Fewer tests may run locally but `testthat::skip*` functions make it transparent what is skipped.  You'll see this pattern in [**arkdb**](https://github.com/ropensci/piggyback/tree/master/tests/testthat). In this case it's an especially good idea to use [mocking](https://books.ropensci.org/http-testing/) with the **webmockr** or *httptest** packages so that you can separate out tests on HTTP content from the actual response:
- Provide enough direction in `CONTRIBUTING` for someone to run the full test suite.  e.g., get a token, fork/clone a repository, etc.  This _is_ the type of effort we expect from reviewers of packages, though a smaller number of contributors will likely go through it. Here's an [**example of more complex contributing directions written in response to a review**](https://github.com/dbca-wa/ruODK/blob/master/CONTRIBUTING.md#test).  A few things would make this easier, for instance, using an environment variable (`CIRCLE_TEST_REPO=http://...`) to define the the URI of the test repo, avoiding the need to adjust the test suite.
- Minimize the testing effort by including set up / tear down the example repository as part of the test suite, thus only requiring the tester to provide keys with adequate scopes. I asked Jenny Bryan, who maintains a lot of packages with complex API interactions and auth setups (e.g. googledrive), about this case, and she strongly recommended a setup/teardown approach for long-term maintainability by both yourself and contributors. I won't require this full level of automation but it's an option in your toolbox.

In any case I'll leave it to reviewers to look at the completeness of the test suite.

Regarding this:

> I don't see why anything should be exposed. The token is stored as a secure env var in the build and nothing will be exposed when a PR is run.

If you allow PRs to use your variables, all one needs to do is slip a `Sys.getenv("GITHUB_TOKEN")` and then a line of code like a `curl` to ship that value elsewhere.  This is why CI systems pretty much all prevent such variables from being exposed to PR builds by default and it's highly recommended to leave it that way.
