### 2020-02-01 #356 @pat-s
Thank Noam, lots of good resources and pointers in there.
It might take me a while to come up with a test suite that fulfills all these requirements.

> for long-term maintainability by both yourself and contributors.

I see the point even though it means a lot of work. Knowing this before might have scared me away even but now I'm in it ;)

> In any case I'll leave it to reviewers to look at the completeness of the test suite.

Happy to wait for reviews until I implemented a more proper testing suite.
