### 2020-02-03 #356 @noamross
To clarify, @pat-s, we _do_ need a test suite that passes locally in order to assign to reviewers, though they may have feedback on completeness or setup.
