### 2020-02-04 169c2f1
fix endpoint leftover in `browse_circle_token()`, fixes #7
