### 2020-02-05 0064d8b
Vignette: Explain better how to set the "user key"
