### 2020-02-05 bd9be3e
update instructions about checkout and ssh keys in vignette
