### 2020-05-26 #356 @noamross
Hello @pat-s, we're starting up review activities so I wanted to check in with you on the status of this package.  Last time we checked in you were developing a test suite in response to my editor's check.  Any developments?  I hope you're well and thank you for your patience!
