### 2020-08-27 e4e2e8c
remove travis and appveyor
