+++ README.Rmd
@@ -3,9 +3,8 @@ output: github_document
 ---
 
 <!-- badges: start -->
-[![Build Status](https://img.shields.io/travis/ropenscilabs/circle/master?label=macOS&logo=travis&style=flat-square)](https://travis-ci.org/ropenscilabs/circle)
+[![tic](https://github.com/ropenscilabs/circle/workflows/tic/badge.svg?branch=master)](https://github.com/ropenscilabs/circle/actions)
 [![CircleCI](https://img.shields.io/circleci/build/gh/ropenscilabs/circle/master?label=Linux&logo=circle&logoColor=green&style=flat-square)](https://circleci.com/gh/ropenscilabs/circle)
-[![AppVeyor build status](https://img.shields.io/appveyor/ci/ropensci/circle?label=Windows&logo=appveyor&style=flat-square)](https://ci.appveyor.com/project/ropensci/circle)
 [![codecov](https://codecov.io/gh/ropenscilabs/circle/branch/master/graph/badge.svg)](https://codecov.io/gh/ropenscilabs/circle)
 [![Lifecycle: maturing](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)
 [![rOpenSci footer](http://ropensci.org/public_images/github_footer.png)](https://ropensci.org)
