### 2020-08-28 57331fe
CI: set R_CIRCLE env var
