### 2020-08-29 8debbba
- print informative message when creating a user key errors with status code 500
