### 2020-08-29 c4194ac
GH_TOKEN_CHECKOUT_KEY -> GH_TOKEN
