+++ .github/workflows/tic.yml
@@ -44,6 +44,8 @@ jobs:
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
       # [Custom env var] Circle CI API key
       R_CIRCLE: ${{ secrets.R_CIRCLE }}
+      # [Custom env var] Circle CI API key
+      CIRCLE_CHECKOUT_KEY: ${{ secrets.CIRCLE_CHECKOUT_KEY }}
 
     steps:
       - uses: actions/checkout@v2.3.1
