+++ .circleci/config.yml
@@ -1,6 +1,9 @@
-version: 2
+## tic Circle CI template: linux
+## revision date: 2020-05-09
+version: 2.1
 jobs:
-  build:
+  r-release:
+    environment:
     docker:
       - image: rocker/verse
     steps:
@@ -9,73 +12,53 @@ jobs:
       # create a unique env var for the cache. Unfortunately normal env vars
       # are not picked up by the cache, therefore this workaround is needed.
       # See https://discuss.circleci.com/t/cannot-use-circle-yml-environment-variables-in-cache-keys/10994/7
-      - run: echo "$(date -d "$D" '+%d')" > /tmp/_tmp_file
+      - run: echo "$(date '+%d-%m')-r-release" > /tmp/_tmp_file
+
       - restore_cache:
           key: R-package-library-{{ checksum "/tmp/_tmp_file" }}
 
       # install deps and check pkg ---------------------------------------------
       - run:
-          name: Install package dependencies
+          name: "[r-release] Install dependencies"
           command: |
-            echo -e 'options(Ncpus = 6, repos = structure(c(CRAN = "https://cloud.r-project.org/")))' > $HOME/.Rprofile
-            R -q -e 'if (!requireNamespace("remotes")) install.packages("remotes")'
+            sudo apt update && sudo apt install -y ccache
+            echo -e 'options(Ncpus = 4, repos = structure(c(CRAN = "https://cloud.r-project.org/")))' > $HOME/.Rprofile
+            mkdir -p $HOME/.ccache && echo -e 'max_size = 5.0G\nsloppiness = include_file_ctime\nhash_dir=false' > $HOME/.ccache/ccache.conf
+            mkdir -p ~/.R && echo -e 'CC=ccache gcc\nCPP=ccache gcc\nCXX=ccache g++\nCXX11=ccache g++\nCXX14=ccache g++\nCXX17=ccache g++\nF77=ccache gfortran\nFC=ccache gfortran' > $HOME/.R/Makevars
+            R -q -e 'install.packages("remotes")'
             R -q -e 'if (getRversion() < "3.2" && !requireNamespace("curl")) install.packages("curl")'
-            R -q -e 'remotes::install_github("ropenscilabs/tic", upgrade = "always"); print(tic::dsl_load()); tic::prepare_all_stages()'
+            R -q -e 'remotes::install_github("ropensci/tic", upgrade = "always"); print(tic::dsl_load()); tic::prepare_all_stages()'
             R -q -e 'tic::before_install()'
             R -q -e 'tic::install()'
+
       - run:
-          name: Check package
+          name: "[r-release] R CMD Check"
           no_output_timeout: 60m
           command: |
             R -q -e 'tic::before_script()'
             R -q -e 'tic::script()'
-            R -q -e 'tic::after_success()'
 
       # save R pkg cache -------------------------------------------------------
       - save_cache:
           key: R-package-library-{{ checksum "/tmp/_tmp_file" }}
           paths:
             - /usr/local/lib/R/site-library
-  deploy:
-    environment:
-      id_rsa: true
-    docker:
-      - image: rocker/verse
-    steps:
-      - checkout
-      - run: echo "$(date -d "$D" '+%d')" > /tmp/_tmp_file
-      - restore_cache:
-          key: R-package-library-{{ checksum "/tmp/_tmp_file" }}
-      - run:
-          name: Deploy Over SSH
-          command: |
-            R -q -e 'if (!requireNamespace("remotes")) install.packages("remotes")'
-            R -q -e 'if (getRversion() < "3.2" && !requireNamespace("curl")) install.packages("curl")'
[…]
