+++ man/ssh_key_helpers.Rd
@@ -1,0 +1,45 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/helpers-github.R
+\name{ssh_key_helpers}
+\alias{ssh_key_helpers}
+\alias{github_add_key}
+\alias{check_admin_repo}
+\alias{get_role_in_repo}
+\title{SSH key helpers}
+\usage{
+github_add_key(
+  pubkey,
+  repo = get_repo(remote),
+  user = get_user(),
+  title = "travis",
+  remote = "origin",
+  check_role = TRUE
+)
+
+check_admin_repo(owner, user, repo)
+
+get_role_in_repo(owner, user, repo)
+}
+\arguments{
+\item{pubkey}{The public key of the SSH key pair}
+
+\item{repo}{`[string]`\cr
+The repository slug to use. Must follow the "`user/repo`" structure.}
+
+\item{user}{The name of the user account}
+
+\item{title}{The title of the key to add}
+
+\item{remote}{`[string]`\cr
+The Github remote which should be used. Defaults to "origin".}
+
+\item{check_role}{Whether to check if the current user has the permissions to
+add a key to the repo. Setting this to `FALSE` makes it possible to add keys
+to other repos than just the one from which the function is called.}
+
+\item{owner}{The owner of the repository}
+}
+\description{
+SSH key helpers
+}
+\keyword{internal}
