+++ NAMESPACE
@@ -3,7 +3,9 @@
 S3method(print,circle_collection)
 S3method(print,circle_job)
 S3method(print,circle_workflow)
+export(auth_github)
 export(browse_circle_token)
+export(check_admin_repo)
 export(circle)
 export(create_checkout_key)
 export(delete_checkout_key)
@@ -13,14 +15,21 @@ export(enable_repo)
 export(get_build_artifacts)
 export(get_checkout_keys)
 export(get_env_vars)
+export(get_owner)
 export(get_pipelines)
+export(get_repo)
+export(get_repo_slug)
+export(get_role_in_repo)
 export(get_user)
+export(github_add_key)
+export(github_info)
 export(github_repo)
 export(has_checkout_key)
 export(list_projects)
 export(new_build)
 export(set_env_var)
 export(use_circle_deploy)
+export(uses_github)
 import(cli)
 import(httr)
 importFrom(cli,cli_text)

[…]
