### 2020-09-03 b34d592
update extract_repo() and get_remote_url()
