+++ R/helpers-github.R
@@ -161,17 +161,19 @@ get_repo_data <- function(repo) {
   return(req)
 }
 
-get_remote_url <- function(path) {
+get_remote_url <- function(path, remote) {
   r <- git2r::repository(path, discover = TRUE)
   remote_names <- git2r::remotes(r)
   if (!length(remote_names)) {
-    stop("Failed to lookup git remotes") # nocov
+    stopc("Failed to lookup git remotes")
   }
-  remote_name <- "origin"
-  if (!("origin" %in% remote_names)) { # nocov start
-    remote_name <- remote_names[1]
-    warning(sprintf("No remote 'origin' found. Using: %s", remote_name))
-  } # nocov end
+  remote_name <- remote
+  if (!(remote_name %in% remote_names)) {
+    stopc(sprintf(
+      "No remote named '%s' found in remotes: '%s'.",
+      remote_name, remote_names
+    ))
+  }
   git2r::remote_url(r, remote_name)
 }
 
@@ -182,12 +184,11 @@ extract_repo <- function(url) {
   match <- regmatches(url, m)[[1]]
 
   if (length(match) == 0) {
-    stop("Unrecognized repo format: ", url) # nocov
+    stopc("Unrecognized repo format: ", url)
   }
 
   paste0(match[2], "/", match[3])
 }
-
 
 # GH auth helpers --------------------------------------------------------------
 
