+++ man/get_circle_user.Rd
@@ -1,10 +1,10 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/general.R
-\name{get_user}
-\alias{get_user}
+\name{get_circle_user}
+\alias{get_circle_user}
 \title{Get Circle CI user}
 \usage{
-get_user()
+get_circle_user()
 }
 \value{
 A list of class `circle_user`.
