+++ NAMESPACE
@@ -14,6 +14,7 @@ export(edit_circle_config)
 export(enable_repo)
 export(get_build_artifacts)
 export(get_checkout_keys)
+export(get_circle_user)
 export(get_env_vars)
 export(get_owner)
 export(get_pipelines)
