### 2020-09-03 b2155fc
add check_private_key_name(), get_public_key()  and encode_private_key_Name()
