+++ R/helpers-github.R
@@ -190,6 +190,40 @@ extract_repo <- function(url) {
   paste0(match[2], "/", match[3])
 }
 
+#' @param key The SSH key pair object
+#' @keywords internal
+#' @name ssh_key_helpers
+#' @export
+get_public_key <- function(key) {
+  as.list(key)$pubkey
+}
+
+#' @param key The SSH key pair object
+#' @keywords internal
+#' @rdname ssh_key_helpers
+#' @export
+encode_private_key <- function(key) {
+  conn <- textConnection(NULL, "w")
+  openssl::write_pem(key, conn, password = NULL)
+  private_key <- textConnectionValue(conn)
+  close(conn)
+
+  private_key <- paste(private_key, collapse = "\n")
+
+  openssl::base64_encode(charToRaw(private_key))
+}
+
+#' @param string String to check
+#' @keywords internal
+#' @rdname ssh_key_helpers
+#' @export
+check_private_key_name <- function(string) {
+  if (grepl("[ ]", string)) {
+    stopc("Name contains whitespaces. Please supply a name without whitespaces.") # nolint
+  }
+  return(invisible(TRUE))
+}
+
 # GH auth helpers --------------------------------------------------------------
 
 #' @title Github API helpers
