+++ NAMESPACE
@@ -6,18 +6,21 @@ S3method(print,circle_workflow)
 export(auth_github)
 export(browse_circle_token)
 export(check_admin_repo)
+export(check_private_key_name)
 export(circle)
 export(create_checkout_key)
 export(delete_checkout_key)
 export(delete_env_var)
 export(edit_circle_config)
 export(enable_repo)
+export(encode_private_key)
 export(get_build_artifacts)
 export(get_checkout_keys)
 export(get_circle_user)
 export(get_env_vars)
 export(get_owner)
 export(get_pipelines)
+export(get_public_key)
 export(get_repo)
 export(get_repo_slug)
 export(get_role_in_repo)
