### 2020-09-03 a03e652
need stopc()
