### 2020-09-03 49f9727
fix tests after get_user_circle() transition
