+++ man/github_info.Rd
@@ -35,4 +35,3 @@ Retrieves metadata about a Git repository from GitHub.
 `github_info()` returns a list as obtained from the GET "/repos/:repo" API.
 }
 \concept{GitHub functions}
-\keyword{internal}
