+++ NAMESPACE
@@ -3,37 +3,23 @@
 S3method(print,circle_collection)
 S3method(print,circle_job)
 S3method(print,circle_workflow)
-export(auth_github)
 export(browse_circle_token)
-export(check_admin_repo)
-export(check_private_key_name)
 export(circle)
 export(create_checkout_key)
 export(delete_checkout_key)
 export(delete_env_var)
 export(edit_circle_config)
 export(enable_repo)
-export(encode_private_key)
 export(get_build_artifacts)
 export(get_checkout_keys)
 export(get_circle_user)
 export(get_env_vars)
-export(get_owner)
 export(get_pipelines)
-export(get_public_key)
-export(get_repo)
-export(get_repo_slug)
-export(get_role_in_repo)
-export(get_user)
-export(github_add_key)
-export(github_info)
-export(github_repo)
 export(has_checkout_key)
 export(list_projects)
 export(new_build)
 export(set_env_var)
 export(use_circle_deploy)
-export(uses_github)
 import(cli)
 import(httr)
 importFrom(cli,cli_text)
