+++ tests/testthat/test-general.R
@@ -1,38 +1,46 @@
 context("general")
 
-withr::with_dir(
-  "travis-testthat",
-  {
-    test_that("get_circle_user() works", {
-      skip_on_cran()
+vcr::use_cassette("get_circle_user()", {
+  test_that("get_circle_user() works", {
+    skip_on_cran()
 
-      expect_s3_class(get_circle_user(), "circle_user")
-    })
+    resp <- get_circle_user()
 
-    test_that("list_projects() works", {
-      skip_on_cran()
+    expect_s3_class(resp, "circle_user")
+    expect_equal(status_code(resp$response), 200)
+  })
+})
 
-      expect_s3_class(
-        list_projects(),
-        "circle_api"
-      )
-    })
+vcr::use_cassette("list_projects()", {
+  test_that("list_projects() works", {
+    skip_on_cran()
 
-    test_that("triggering a new build works", {
-      skip_on_cran()
+    resp <- list_projects()
 
-      expect_s3_class(
-        new_build(),
-        "circle_api"
-      )
-    })
+    expect_equal(status_code(resp$response), 200)
+    expect_gte(length(resp$content), 1)
+    expect_s3_class(resp, "circle_api")
+  })
+})
 
-    test_that("checking the existence of checkout keys works", {
-      skip_on_cran()
+vcr::use_cassette("new_build()", {
+  test_that("triggering a new build works", {
+    skip_on_cran()
 
-      expect_true(
-        has_checkout_key()
-      )
-    })
-  }
-)
+    resp <- new_build()
+
+    expect_equal(status_code(resp$response), 201)
+    expect_match(resp[["content"]][["state"]], "pending")
+    expect_s3_class(new_build(), "circle_api")
+  })
+})
+
+vcr::use_cassette("has_checkout_key()", {
+  test_that("checking the existence of checkout keys works", {
+    skip_on_cran()
+
+    resp <- has_checkout_key()
+
+    expect_true(resp)
+  })
+})
