+++ tests/fixtures/list_projects().yml
@@ -1,0 +1,41 @@
+http_interactions:
+- request:
+    method: get
+    uri: https://circleci.com/api/v1.1/projects?circle-token=R_CIRCLE
+    body:
+      encoding: ''
+      string: ''
+    headers:
+      Content-Type: ''
+      Accept: application/json
+  response:
+    status:
+      status_code: 200
+      category: Success
+      reason: OK
+      message: 'Success: (200) OK'
+    headers:
+      access-control-allow-origin: '*'
+      content-encoding: gzip
+      content-type: application/json;charset=utf-8
+      date: Fri, 04 Sep 2020 19:27:47 GMT
+      server: nginx
+      set-cookie: ring-session=LZorB2VwjOHv3nSO4A9UvHmx4A%2FLIXyq5ibD2sNUPYvCn8uv2%2By2PCLPecwJ3q54Ho5N1u%2BLsZRArVr43eGENBLuVXCHbcIXtI1VPM2bYXrbMRzy51I0dSLapmqkR9f%2FzCGzZR2cjoAnKcQ7bOcY%2Bx%2BT%2FN7oFf8%2FksijbnXqGFKMSSwEBkJl3HN5o8cop8ghTI%2BsXUppmXLn9YoJAHDh5L%2BzY9cNWSVd5ECw1ht8y2k%3D--laAKiWcZ6gAGcvqqyUu7iz9RBD%2BaRTSK6xnHQtesIL4%3D;Path=/;HttpOnly;Expires=Sat,
+        04 Sep 2021 13:58:35 +0000;Max-Age=1209600;Secure
+      strict-transport-security: max-age=15724800
+      vary: Origin
+      x-circleci-identity: circle-www-api-v1-7778bddccf-nhp8q
+      x-frame-options: DENY
+      x-ratelimit-limit: '151'
+      x-ratelimit-remaining: '150'
+      x-ratelimit-reset: '0'
+      x-request-id: b92647ca-f0e4-4cd3-922b-dc475519b0f0
+      x-route: /api/v1.1/projects
+      transfer-encoding: chunked
+      connection: keep-alive
+    body:
+      encoding: UTF-8
+      file: no
[…]
