+++ tests/fixtures/has_checkout_key().yml
@@ -1,0 +1,157 @@
+http_interactions:
+- request:
+    method: get
+    uri: https://api.github.com/repos/ropenscilabs/circle
+    body:
+      encoding: ''
+      string: ''
+    headers:
+      User-Agent: https://github.com/r-lib/gh
+      Accept: application/vnd.github.v3+json
+      Authorization: token 23882cdaacc22ebafa63beafa7d1e92ce009e3c4
+  response:
+    status:
+      status_code: 200
+      category: Success
+      reason: OK
+      message: 'Success: (200) OK'
+    headers:
+      access-control-allow-origin: '*'
+      access-control-expose-headers: ETag, Link, Location, Retry-After, X-GitHub-OTP,
+        X-RateLimit-Limit, X-RateLimit-Remaining, X-RateLimit-Used, X-RateLimit-Reset,
+        X-OAuth-Scopes, X-Accepted-OAuth-Scopes, X-Poll-Interval, X-GitHub-Media-Type,
+        Deprecation, Sunset
+      cache-control: private, max-age=60, s-maxage=60
+      content-encoding: gzip
+      content-security-policy: default-src 'none'
+      content-type: application/json; charset=utf-8
+      date: Fri, 04 Sep 2020 19:27:50 GMT
+      etag: W/"35c7938e9948146693e9ac7f51b77100"
+      last-modified: Fri, 04 Sep 2020 11:36:52 GMT
+      referrer-policy: origin-when-cross-origin, strict-origin-when-cross-origin
+      server: GitHub.com
+      status: 200 OK
+      strict-transport-security: max-age=31536000; includeSubdomains; preload
+      transfer-encoding: chunked
+      vary:
+      - Accept, Authorization, Cookie, X-GitHub-OTP
+      - Accept-Encoding, Accept, X-Requested-With
+      - Accept-Encoding
+      x-accepted-oauth-scopes: repo
+      x-content-type-options: nosniff
+      x-frame-options: deny
+      x-github-media-type: github.v3; format=json
+      x-github-request-id: 2603:BB64:C071A4D:E828996:5F529535
+      x-oauth-scopes: gist, repo
+      x-ratelimit-limit: '5000'
+      x-ratelimit-remaining: '4949'
+      x-ratelimit-reset: '1599250807'
+      x-ratelimit-used: '51'
+      x-xss-protection: 1; mode=block
+    body:
+      encoding: UTF-8
+      file: no
[…]
