+++ DESCRIPTION
@@ -31,6 +31,7 @@ Suggests:
     testthat,
     usethis,
     utils,
+    vcr,
     withr
 VignetteBuilder: 
     knitr

[…]
