### 2020-09-23 933e9ad
try new filtering approach
