+++ tests/fixtures/has_checkout_key().yml
@@ -8,7 +8,7 @@ http_interactions:
     headers:
       User-Agent: https://github.com/r-lib/gh
       Accept: application/vnd.github.v3+json
-      Authorization: token 23882cdaacc22ebafa63beafa7d1e92ce009e3c4
+      Authorization: token <<<PAT_GITHUB>>>
   response:
     status:
       status_code: 200
@@ -25,8 +25,8 @@ http_interactions:
       content-encoding: gzip
       content-security-policy: default-src 'none'
       content-type: application/json; charset=utf-8
-      date: Fri, 04 Sep 2020 19:27:50 GMT
-      etag: W/"35c7938e9948146693e9ac7f51b77100"
+      date: Wed, 23 Sep 2020 20:13:46 GMT
+      etag: W/"555425e64fac704b3e3218f239d390222f9ee4c1a6607db973a23360149ecb63"
       last-modified: Fri, 04 Sep 2020 11:36:52 GMT
       referrer-policy: origin-when-cross-origin, strict-origin-when-cross-origin
       server: GitHub.com
@@ -41,20 +41,20 @@ http_interactions:
       x-content-type-options: nosniff
       x-frame-options: deny
       x-github-media-type: github.v3; format=json
-      x-github-request-id: 2603:BB64:C071A4D:E828996:5F529535
-      x-oauth-scopes: gist, repo
+      x-github-request-id: E2B3:131CB:746F942:8B27603:5F6BAC79
+      x-oauth-scopes: repo
       x-ratelimit-limit: '5000'
-      x-ratelimit-remaining: '4949'
-      x-ratelimit-reset: '1599250807'
-      x-ratelimit-used: '51'
+      x-ratelimit-remaining: '4945'
+      x-ratelimit-reset: '1600895252'
+      x-ratelimit-used: '55'
       x-xss-protection: 1; mode=block
     body:
       encoding: UTF-8
       file: no
       string: '{"id":214703960,"node_id":"MDEwOlJlcG9zaXRvcnkyMTQ3MDM5NjA=","name":"circle","full_name":"ropenscilabs/circle","private":false,"owner":{"login":"ropenscilabs","id":16314060,"node_id":"MDEyOk9yZ2FuaXphdGlvbjE2MzE0MDYw","avatar_url":"https://avatars3.githubusercontent.com/u/16314060?v=4","gravatar_id":"","url":"https://api.github.com/users/ropenscilabs","html_url":"https://github.com/ropenscilabs","followers_url":"https://api.github.com/users/ropenscilabs/followers","following_url":"https://api.github.com/users/ropenscilabs/following{/other_user}","gists_url":"https://api.github.com/users/ropenscilabs/gists{/gist_id}","starred_url":"https://api.github.com/users/ropenscilabs/starred{/owner}{/repo}","subscriptions_url":"https://api.github.com/users/ropenscilabs/subscriptions","organizations_url":"https://api.github.com/users/ropenscilabs/orgs","repos_url":"https://api.github.com/users/ropenscilabs/repos","events_url":"https://api.github.com/users/ropenscilabs/events{/privacy}","received_events_url":"https://api.github.com/users/ropenscilabs/received_events","type":"Organization","site_admin":false},"html_url":"https://github.com/ropenscilabs/circle","description":"R
[…]
