+++ tests/fixtures/new_build().yml
@@ -25,8 +25,8 @@ http_interactions:
       content-encoding: gzip
       content-security-policy: default-src 'none'
       content-type: application/json; charset=utf-8
-      date: Wed, 23 Sep 2020 20:13:44 GMT
-      etag: W/"555425e64fac704b3e3218f239d390222f9ee4c1a6607db973a23360149ecb63"
+      date: Wed, 23 Sep 2020 21:04:09 GMT
+      etag: W/"afdacf577a9ddd33bbb32ea4326811a69c83694fce8b850d09af2f846e3990f7"
       last-modified: Fri, 04 Sep 2020 11:36:52 GMT
       referrer-policy: origin-when-cross-origin, strict-origin-when-cross-origin
       server: GitHub.com
@@ -41,20 +41,20 @@ http_interactions:
       x-content-type-options: nosniff
       x-frame-options: deny
       x-github-media-type: github.v3; format=json
-      x-github-request-id: E2B3:131CB:746F797:8B2745E:5F6BAC78
+      x-github-request-id: F2EF:F313:C76AE68:EF63E3D:5F6BB849
       x-oauth-scopes: repo
       x-ratelimit-limit: '5000'
-      x-ratelimit-remaining: '4952'
+      x-ratelimit-remaining: '4871'
       x-ratelimit-reset: '1600895252'
-      x-ratelimit-used: '48'
+      x-ratelimit-used: '129'
       x-xss-protection: 1; mode=block
     body:
       encoding: UTF-8
       file: no
       string: '{"id":214703960,"node_id":"MDEwOlJlcG9zaXRvcnkyMTQ3MDM5NjA=","name":"circle","full_name":"ropenscilabs/circle","private":false,"owner":{"login":"ropenscilabs","id":16314060,"node_id":"MDEyOk9yZ2FuaXphdGlvbjE2MzE0MDYw","avatar_url":"https://avatars3.githubusercontent.com/u/16314060?v=4","gravatar_id":"","url":"https://api.github.com/users/ropenscilabs","html_url":"https://github.com/ropenscilabs","followers_url":"https://api.github.com/users/ropenscilabs/followers","following_url":"https://api.github.com/users/ropenscilabs/following{/other_user}","gists_url":"https://api.github.com/users/ropenscilabs/gists{/gist_id}","starred_url":"https://api.github.com/users/ropenscilabs/starred{/owner}{/repo}","subscriptions_url":"https://api.github.com/users/ropenscilabs/subscriptions","organizations_url":"https://api.github.com/users/ropenscilabs/orgs","repos_url":"https://api.github.com/users/ropenscilabs/repos","events_url":"https://api.github.com/users/ropenscilabs/events{/privacy}","received_events_url":"https://api.github.com/users/ropenscilabs/received_events","type":"Organization","site_admin":false},"html_url":"https://github.com/ropenscilabs/circle","description":"R
[…]
