### 2020-09-24 #356 @pat-s
@noamross

Just wanted to inform you that I've started to incorporate {vcr} for http testing.
Still facing issues and lacking time but this project is not stalled (anymore) :)
