+++ tests/testthat/helpers.R
@@ -1,3 +1,3 @@
-if (!dir.exists("travis-testthat")) {
-  system("git clone https://github.com/pat-s/travis-testthat.git")
-}
+# if (!dir.exists("travis-testthat")) {
+#   system("git clone https://github.com/pat-s/travis-testthat.git")
+# }
