+++ man/github_helpers.Rd
@@ -22,6 +22,12 @@ get_repo_slug(remote = "origin")
 \arguments{
 \item{remote}{`[string]`\cr
 The Github remote which should be used. Defaults to "origin".}
+
+\item{.token}{`[character]`\cr
+Authentication token. Defaults to GITHUB_PAT or GITHUB_TOKEN environment
+variables, in this order if any is set. See gh_token() if you need more
+flexibility, e.g. different tokens for different GitHub Enterprise
+deployments.}
 }
 \description{
 - `auth_github()`: Creates a `GITHUB_TOKEN` and asks to store it in your
