+++ tests/testthat/test-auth.R
@@ -1,13 +1,10 @@
-# withr::with_dir(
-#   "travis-testthat",
-#   {
-#     test_that("Circle enable works", {
-#       skip_on_cran()
-#
-#       # enable
-#       foo <- suppressMessages(enable_repo())
-#
-#       expect_true(foo)
-#     })
-#   }
-# )
+vcr::use_cassette("enable_repo()", {
+  test_that("enable_repo() works", {
+    skip_on_cran()
+
+    # enable
+    foo <- suppressMessages(enable_repo())
+
+    expect_true(foo)
+  })
+})
