+++ tests/testthat/test-general.R
@@ -4,6 +4,9 @@ vcr::use_cassette("get_circle_user()", {
   test_that("get_circle_user() works", {
     skip_on_cran()
 
+    # 'repo' and 'user' need to be set explicitly because `github_info()` will
+    # fail to lookup the git repo when running code coverage
+
     resp <- get_circle_user()
 
     expect_s3_class(resp, "circle_user")
@@ -15,7 +18,9 @@ vcr::use_cassette("list_projects()", {
   test_that("list_projects() works", {
     skip_on_cran()
 
-    resp <- list_projects()
+    # 'repo' and 'user' need to be set explicitly because `github_info()` will
+    # fail to lookup the git repo when running code coverage
+    resp <- list_projects(repo = "circle", user = "ropenscilabs")
 
     expect_equal(status_code(resp$response), 200)
     expect_gte(length(resp$content), 1)
@@ -27,14 +32,18 @@ vcr::use_cassette("new_build()", {
   test_that("triggering a new build works", {
     skip_on_cran()
 
+    # 'repo' and 'user' need to be set explicitly because `github_info()` will
+    # fail to lookup the git repo when running code coverage
     resp <- new_build(
-      repo = github_info(.token = Sys.getenv("PAT_GITHUB"))$name,
-      user = github_info(.token = Sys.getenv("PAT_GITHUB"))$owner$login
+      repo = "circle", user = "ropenscilabs"
     )
 
     expect_equal(status_code(resp$response), 201)
     expect_match(resp[["content"]][["state"]], "pending")
-    expect_s3_class(new_build(), "circle_api")
+    expect_s3_class(
+      new_build(repo = "circle", user = "ropenscilabs"),
+      "circle_api"
+    )
   })
 })
 
@@ -42,7 +51,9 @@ vcr::use_cassette("has_checkout_key()", {
   test_that("checking the existence of checkout keys works", {
     skip_on_cran()
 
-    resp <- has_checkout_key()
+    # 'repo' and 'user' need to be set explicitly because `github_info()` will
+    # fail to lookup the git repo when running code coverage
+    resp <- has_checkout_key(repo = "circle", user = "ropenscilabs")
 
     expect_true(resp)
   })
