+++ R/helpers-github.R
@@ -11,91 +11,91 @@
 #' to other repos than just the one from which the function is called.
 #' @template remote
 #' @name ssh_key_helpers
-github_add_key <- function(pubkey,
-                           repo = get_repo(remote),
-                           user = get_user(),
-                           title = "travis",
-                           remote = "origin",
-                           check_role = TRUE) {
+# github_add_key <- function(pubkey,
+#                            repo = get_repo(remote),
+#                            user = get_user(),
+#                            title = "travis",
+#                            remote = "origin",
+#                            check_role = TRUE) {
 
-  if (inherits(pubkey, "key")) {
-    pubkey <- as.list(pubkey)$pubkey
-  }
-  if (!inherits(pubkey, "pubkey")) {
-    stop("`pubkey` must be an RSA/EC public key")
-  }
+#   if (inherits(pubkey, "key")) {
+#     pubkey <- as.list(pubkey)$pubkey
+#   }
+#   if (!inherits(pubkey, "pubkey")) {
+#     stop("`pubkey` must be an RSA/EC public key")
+#   }
 
-  if (check_role) {
-    # check if we have enough rights to add a key
-    check_admin_repo(
-      owner = get_owner(remote = remote),
-      user = user,
-      repo = repo
-    )
-  }
-  key_data <- create_key_data(pubkey, title)
+#   if (check_role) {
+#     # check if we have enough rights to add a key
+#     check_admin_repo(
+#       owner = get_owner(remote = remote),
+#       user = user,
+#       repo = repo
+#     )
+#   }
+#   key_data <- create_key_data(pubkey, title)
 
-  # add public key to repo deploy keys on GitHub
-  ret <- add_key(key_data,
-    owner = get_owner(remote = remote),
-    project = repo
-  )
+#   # add public key to repo deploy keys on GitHub
+#   ret <- add_key(key_data,
+#     owner = get_owner(remote = remote),
+#     project = repo
+#   )
 
-  cli::cli_alert_success("{.fun github_add_key}: Added a public deploy key to
-    GitHub for repo {.field {get_owner(remote)}/{repo}}.", wrap = TRUE)
+#   cli::cli_alert_success("{.fun github_add_key}: Added a public deploy key to
+#     GitHub for repo {.field {get_owner(remote)}/{repo}}.", wrap = TRUE)
 
-  invisible(ret)
-}
-
-#' @param owner The owner of the repository
-#' @param user The name of the user account
-#' @template repo
-#' @rdname ssh_key_helpers
-check_admin_repo <- function(owner, user, repo) {
-  role_in_repo <- get_role_in_repo(owner, user, repo)
-  if (role_in_repo != "admin") {
-    stop(
-      "Must have role 'admin' to add deploy key to repo ",
-      repo, ", not '", role_in_repo, "'."
-    )
-  }
-}
-
-add_key <- function(key_data, owner, project) {
-
-  resp <- gh::gh("POST /repos/:owner/:repo/keys",
-    owner = owner, repo = project,
-    title = key_data$title,
-    key = key_data$key, read_only = key_data$read_only
-  )
-
-  invisible(resp)
-}
+#   invisible(ret)
+# }
 
 #' @param owner The owner of the repository
 #' @param user The name of the user account
 #' @template repo
 #' @rdname ssh_key_helpers
-get_role_in_repo <- function(owner, user, repo) {
+# check_admin_repo <- function(owner, user, repo) {
+#   role_in_repo <- get_role_in_repo(owner, user, repo)
+#   if (role_in_repo != "admin") {
+#     stop(
[…]
