+++ man/github_info.Rd
@@ -1,28 +1,14 @@
 % Generated by roxygen2: do not edit by hand
 % Please edit documentation in R/helpers-github.R
-\name{github_repo}
-\alias{github_repo}
+\name{github_info}
 \alias{github_info}
-\alias{uses_github}
 \title{Github information}
 \usage{
-github_repo(
-  path = usethis::proj_get(),
-  info = github_info(path, remote = remote),
-  remote = "origin"
-)
-
-github_info(path = usethis::proj_get(), remote = "origin", .token = NULL)
-
-uses_github(path = usethis::proj_get())
+github_info(path = ".", remote = "origin", .token = NULL)
 }
 \arguments{
 \item{path}{`[string]`\cr
 The path to a GitHub-enabled Git repository (or a subdirectory thereof).}
-
-\item{info}{`[list]`\cr
-GitHub information for the repository, by default obtained through
-[github_info()].}
 
 \item{remote}{`[string]`\cr
 The Github remote which should be used. Defaults to "origin".}
@@ -34,8 +20,6 @@ flexibility, e.g. different tokens for different GitHub Enterprise
 deployments.}
 }
 \description{
-`github_repo()` returns the true repository name as string.
-
 Retrieves metadata about a Git repository from GitHub.
 
 `github_info()` returns a list as obtained from the GET "/repos/:repo" API.
