+++ man/ssh_key_helpers.Rd
@@ -1,57 +1,0 @@
-% Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/helpers-github.R
-\name{ssh_key_helpers}
-\alias{ssh_key_helpers}
-\alias{github_add_key}
-\alias{check_admin_repo}
-\alias{get_role_in_repo}
-\alias{get_public_key}
-\alias{encode_private_key}
-\alias{check_private_key_name}
-\title{SSH key helpers}
-\usage{
-github_add_key(
-  pubkey,
-  repo = get_repo(remote),
-  user = get_user(),
-  title = "travis",
-  remote = "origin",
-  check_role = TRUE
-)
-
-check_admin_repo(owner, user, repo)
-
-get_role_in_repo(owner, user, repo)
-
-get_public_key(key)
-
-encode_private_key(key)
-
-check_private_key_name(string)
-}
-\arguments{
-\item{pubkey}{The public key of the SSH key pair}
-
-\item{repo}{`[string]`\cr
-The repository slug to use. Must follow the "`user/repo`" structure.}
-
-\item{user}{The name of the user account}
-
-\item{title}{The title of the key to add}
-
-\item{remote}{`[string]`\cr
-The Github remote which should be used. Defaults to "origin".}
-
-\item{check_role}{Whether to check if the current user has the permissions to
-add a key to the repo. Setting this to `FALSE` makes it possible to add keys
-to other repos than just the one from which the function is called.}
-
-\item{owner}{The owner of the repository}
-
-\item{key}{The SSH key pair object}
-
-\item{string}{String to check}
-}
-\description{
-SSH key helpers
-}
