### 2020-11-03 9381006
get GH Token on CI conditionally
