+++ .pre-commit-config.yaml
@@ -1,6 +1,6 @@
 repos:
--   repo: https://github.com/lorenzwalthert/pre-commit-hooks
-    rev: v0.1.2
+-   repo: https://github.com/lorenzwalthert/precommit
+    rev: v0.1.3
     hooks:
     -   id: style-files
         args: [--style_fun=tidyverse_style]
@@ -9,7 +9,7 @@ repos:
         args: [--warn_only]
     -   id: no-browser-statement
     -   id: readme-rmd-rendered
-    #-   id: roxygenize
+    -   id: roxygenize
     -   id: codemeta-description-updated
     -   id: use-tidy-description
 -   repo: https://github.com/lorenzwalthert/precommit-markdown-link-check
