### 2020-11-03 2fcfc53
pass gh token to create_from_github explicitly
