+++ DESCRIPTION
@@ -7,10 +7,10 @@ Authors@R:
            role = c("aut", "cre"),
            email = "patrick.schratz@gmail.com",
            comment = c(ORCID = "0000-0003-0748-6624"))
-Description: Tools for interacting with the 'Circle CI' API.
-    Besides executing common tasks such as querying build logs and
-    restarting builds, this package also helps setting up permissions to
-    deploy from builds.
+Description: Tools for interacting with the 'Circle CI' API.  Besides
+    executing common tasks such as querying build logs and restarting
+    builds, this package also helps setting up permissions to deploy from
+    builds.
 License: GPL-3
 URL: https://ropenscilabs.github.io/circle,
     https://github.com/ropenscilabs/circle
@@ -28,13 +28,15 @@ Suggests:
     openssl,
     purrr,
     rmarkdown,
-    testthat,
+    testthat (>= 3.0.0),
     usethis,
     utils,
     vcr,
     withr
 VignetteBuilder:
     knitr
+Config/testthat/edition: 3
+Config/testthat/parallel: false
 Encoding: UTF-8
 LazyData: true
 RoxygenNote: 7.1.1
