+++ README.md
@@ -1,4 +1,3 @@
-
 <!-- badges: start -->
 
 [![tic](https://github.com/ropenscilabs/circle/workflows/tic/badge.svg?branch=master)](https://github.com/ropenscilabs/circle/actions)
@@ -6,12 +5,10 @@
 [![codecov](https://codecov.io/gh/ropenscilabs/circle/branch/master/graph/badge.svg)](https://codecov.io/gh/ropenscilabs/circle)
 [![Lifecycle:
 maturing](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://www.tidyverse.org/lifecycle/#maturing)
-[![rOpenSci
-footer](http://ropensci.org/public_images/github_footer.png)](https://ropensci.org)
+
 <!-- badges: end -->
 
-circle
-======
+# circle
 
 R client package for the Circle CI REST API
 
@@ -24,8 +21,7 @@ process for build deployments via `use_circle_deploy()`.
 Please see the related [{tic}](https://github.com/ropensci/tic) package
 for such functionality.
 
-API versions
-------------
+## API versions
 
 All functionality uses the Circle CI [API
 v2](https://github.com/CircleCI-Public/api-preview-docs) which follows
@@ -42,33 +38,33 @@ If you want to get more information, have a look at the [document
 explaining changes between v1.1 and
 v2](https://github.com/CircleCI-Public/api-preview-docs/blob/master/docs/api-changes.md#endpoints-likely-being-removed-in-api-v2-still-available-in-v11-for-now).
 
-Installation
-------------
+## Installation
 
 Development Version:
 
-    remotes::install_github("ropenscilabs/circle")
+```r
+remotes::install_github("ropenscilabs/circle")
+```
 
-Get Started
------------
+## Get Started
 
 See the [Getting
 Started](https://ropenscilabs.github.io/circle/articles/circle.html)
 vignette for an introduction.
 
-Note to developers
-------------------
+## Note to developers
 
 This packages relies on using API keys to test its functionality. Please
 have a look at section [“Testing the
 package”](https://github.com/ropenscilabs/circle/blob/master/.github/CONTRIBUTING.md#testing-the-package)
 for more information.
 
-Acknowledgments
-===============
+# Acknowledgments
 
 This package was inspired by the work of [Thomas J.
 Leeper](https://github.com/leeper) on the (discontinued)
 [cloudyr/circleci](https://github.com/cloudyr/circleci) package and by
 the [ropenscilabs/travis](https://github.com/ropenscilabs/travis)
 package.
+
+[![rOpenSci footer](http://ropensci.org/public_images/github_footer.png)](https://ropensci.org)
