+++ tests/fixtures/enable_repo().yml
@@ -1,119 +1,5 @@
 http_interactions:
 - request:
-    method: get
-    uri: https://api.github.com/repos/ropenscilabs/circle
-    body:
-      encoding: ''
-      string: ''
-    headers:
-      User-Agent: https://github.com/r-lib/gh
-      Accept: application/vnd.github.v3+json
-      Authorization: token <<<PAT_GITHUB>>>
-  response:
-    status:
-      status_code: 200
-      category: Success
-      reason: OK
-      message: 'Success: (200) OK'
-    headers:
-      access-control-allow-origin: '*'
-      access-control-expose-headers: ETag, Link, Location, Retry-After, X-GitHub-OTP,
-        X-RateLimit-Limit, X-RateLimit-Remaining, X-RateLimit-Used, X-RateLimit-Reset,
-        X-OAuth-Scopes, X-Accepted-OAuth-Scopes, X-Poll-Interval, X-GitHub-Media-Type,
-        Deprecation, Sunset
-      cache-control: private, max-age=60, s-maxage=60
-      content-encoding: gzip
-      content-security-policy: default-src 'none'
-      content-type: application/json; charset=utf-8
-      date: Tue, 06 Oct 2020 19:30:37 GMT
-      etag: W/"15f6923db9759a830f1515a59f12e5308f981ec8de3b4a033967d0d39b028263"
-      last-modified: Fri, 04 Sep 2020 11:36:52 GMT
-      referrer-policy: origin-when-cross-origin, strict-origin-when-cross-origin
-      server: GitHub.com
-      status: 200 OK
-      strict-transport-security: max-age=31536000; includeSubdomains; preload
-      transfer-encoding: chunked
-      vary:
-      - Accept, Authorization, Cookie, X-GitHub-OTP
-      - Accept-Encoding, Accept, X-Requested-With
-      - Accept-Encoding
-      x-accepted-oauth-scopes: repo
-      x-content-type-options: nosniff
-      x-frame-options: deny
-      x-github-media-type: github.v3; format=json
-      x-github-request-id: E12F:FB11:36FE5:4141E:5F7CC5DD
-      x-oauth-scopes: repo
-      x-ratelimit-limit: '5000'
-      x-ratelimit-remaining: '4923'
-      x-ratelimit-reset: '1602013530'
-      x-ratelimit-used: '77'
-      x-xss-protection: 1; mode=block
-    body:
-      encoding: UTF-8
-      file: no
[…]
