### 2020-11-03 4ddd4ba
enable flexible local testing
