+++ tests/testthat/test-general.R
@@ -18,7 +18,10 @@ vcr::use_cassette("list_projects()", {
 
     # 'repo' and 'user' need to be set explicitly because `github_info()` will
     # fail to lookup the git repo when running code coverage
-    resp <- list_projects(repo = "circle", user = "ropenscilabs")
+    resp <- list_projects(
+      repo = Sys.getenv("CIRCLE_REPO"),
+      user = Sys.getenv("CIRCLE_OWNER")
+    )
 
     expect_equal(status_code(resp$response), 200)
     expect_gte(length(resp$content), 1)
@@ -39,7 +42,10 @@ vcr::use_cassette("new_build()", {
     expect_equal(status_code(resp$response), 201)
     expect_match(resp[["content"]][["state"]], "pending")
     expect_s3_class(
-      new_build(repo = "circle", user = "ropenscilabs"),
+      new_build(
+        repo = Sys.getenv("CIRCLE_REPO"),
+        user = Sys.getenv("CIRCLE_OWNER")
+      ),
       "circle_api"
     )
   })
@@ -51,7 +57,10 @@ vcr::use_cassette("has_checkout_key()", {
 
     # 'repo' and 'user' need to be set explicitly because `github_info()` will
     # fail to lookup the git repo when running code coverage
-    resp <- has_checkout_key(repo = "circle", user = "ropenscilabs")
+    resp <- has_checkout_key(
+      repo = Sys.getenv("CIRCLE_REPO"),
+      user = Sys.getenv("CIRCLE_OWNER")
+    )
 
     expect_true(resp)
   })
