+++ tests/testthat/test-checkout-keys.R
@@ -8,7 +8,8 @@ vcr::use_cassette("create_checkout_key()", {
     expect_s3_class(
       create_checkout_key(
         type = "user-key",
-        repo = "circle", user = "ropenscilabs"
+        repo = Sys.getenv("CIRCLE_REPO"),
+        user = Sys.getenv("CIRCLE_OWNER")
       ),
       "circle_api"
     )
@@ -22,7 +23,10 @@ vcr::use_cassette("delete_checkout_key()", {
     # 'repo' and 'user' need to be set explicitly because `github_info()` will
     # fail to lookup the git repo when running code coverage
 
-    keys <- get_checkout_keys(repo = "circle", user = "ropenscilabs")
+    keys <- get_checkout_keys(
+      repo = Sys.getenv("CIRCLE_REPO"),
+      user = Sys.getenv("CIRCLE_OWNER")
+    )
     expect_s3_class(keys, "circle_api")
 
     fp <- content(keys$response)$items[[1]]$fingerprint
