+++ tests/testthat/test-auth.R
@@ -4,7 +4,10 @@ vcr::use_cassette("enable_repo()", record = "new_episodes", {
 
     # 'repo' and 'user' need to be set explicitly because `github_info()` will
     # fail to lookup the git repo when running code coverage
-    foo <- suppressMessages(enable_repo(repo = "circle", user = "ropenscilabs"))
+    foo <- suppressMessages(enable_repo(
+      repo = Sys.getenv("CIRCLE_REPO"),
+      user = Sys.getenv("CIRCLE_OWNER")
+    ))
 
     expect_true(foo)
   })
