+++ tests/testthat/helper-circle.R
@@ -9,3 +9,7 @@ invisible(vcr::vcr_configure(
   log_opts = list(file = "console")
 ))
 vcr::check_cassette_names()
+
+# set repo with access right on Circle CI here for local testing
+Sys.setenv("CIRCLE_REPO" = "circle")
+Sys.setenv("CIRCLE_OWNER" = "ropenscilabs")
