+++ tests/testthat/test-helpers-github.R
@@ -1,11 +1,15 @@
 test_that("github helper functions work", {
   unlink(paste0(tempdir(), "/circle"), recursive = TRUE)
   gert::git_clone(
-    "https://github.com/ropenscilabs/circle.git",
-    paste0(tempdir(), "/circle")
+    sprintf(
+      "https://github.com/%s/%s.git",
+      Sys.getenv("CIRCLE_OWNER"),
+      Sys.getenv("CIRCLE_REPO")
+    ),
+    paste0(tempdir(), sprintf("/%s", Sys.getenv("CIRCLE_REPO")))
   )
 
-  withr::with_dir(paste0(tempdir(), "/circle"), {
+  withr::with_dir(paste0(tempdir(), sprintf("/%s", Sys.getenv("CIRCLE_REPO"))), {
 
     # github_info() ------------------------------------------------------------
     if (Sys.getenv("CI") != "") {
