+++ tests/testthat/test-1-builds.R
@@ -6,7 +6,12 @@ vcr::use_cassette("get_jobs()", {
     # fail to lookup the git repo when running code coverage
 
     # calls `get_workflows()` and `get_pipelines()` internally
-    out <- suppressMessages(get_jobs(repo = "circle", user = "ropenscilabs"))
+    out <- suppressMessages(
+      get_jobs(
+        repo = Sys.getenv("CIRCLE_REPO"),
+        user = Sys.getenv("CIRCLE_OWNER")
+      )
+    )
 
     expect_s3_class(out, "circle_collection")
   })
@@ -19,7 +24,10 @@ vcr::use_cassette("get_build_artifacts()", {
     # 'repo' and 'user' need to be set explicitly because `github_info()` will
     # fail to lookup the git repo when running code coverage
     expect_s3_class(
-      get_build_artifacts(repo = "circle", user = "ropenscilabs"),
+      get_build_artifacts(
+        repo = Sys.getenv("CIRCLE_REPO"),
+        user = Sys.getenv("CIRCLE_OWNER")
+      ),
       "circle_api"
     )
   })
@@ -32,8 +40,11 @@ vcr::use_cassette("retry_workflow()", {
     # 'repo' and 'user' need to be set explicitly because `github_info()` will
     # fail to lookup the git repo when running code coverage
     workflow_id <- suppressMessages(
-      get_workflows(repo = "circle", user = "ropenscilabs")
-    )[[10]]$id
+      get_workflows(
+        repo = Sys.getenv("CIRCLE_REPO"),
+        user = Sys.getenv("CIRCLE_OWNER")
+      )[[10]]$id
+    )
     expect_s3_class(
       retry_workflow(workflow_id),
       "circle_api"
