+++ vignettes/circle.Rmd
@@ -24,13 +24,13 @@ There are two ways you can store the API key:
 
   - Via environment variable `R_CIRCLE` in `~/.Renviron` (**recommended**)
   - In `~/.circleci/cli.yml` via the following pattern
-   
+
    ```yml
    host: https://circleci.com
    endpoint: graphql-unstable
    token: <token>
    ```
-   
+
 In addition, {circle} scrapes information about the current repository from Github.
 To be able to do this, a `GITHUB_TOKEN` is needed.
 You can create one using `usethis::browse_github_token()` if you haven't done so yet.
@@ -43,7 +43,7 @@ The easiest way to enable deployment from Circle CI builds to Github repositorie
 circle::use_circle_deploy()
 ✖ No 'user-key' found.
 ──────────────────────────────────────────────────────────────────────────────────────────────────
-✔ Added a 'user key' to project '<repo-slug>' on Circle CI. 
+✔ Added a 'user key' to project '<repo-slug>' on Circle CI.
 This enables deployment from builds.
 ```
 
@@ -61,32 +61,22 @@ If you haven't set a "user-key" yet, there will be a button on this page "Add Us
 
 ## Deployment keys
 
-There two different types of keys on Circle CI: 
+There two different types of keys on Circle CI:
 
-- Checkout keys
+- Deploy key
 
-- SSH keys
+- User key
 
-### Checkout keys
-
-Checkout keys are used to checkout your repository so that the build is starting.
+"Deploy keys" are used to checkout your repository so that the build is starting.
 These kind of keys how only "read" permissions but are not allow to "write" to your repository.
-If you've connected Circle CI to Github already, you will have a "deploy key" stored in every repository to be able to checkout the code.
+If you have connected Circle CI to Github already, you will have a "deploy key" stored in every repository to be able to checkout the code.
 The name "deploy" key here is a bit misleading since they **cannot** be used for deploying from builds.
 
-To enable deployment, a "user-key" is needed which has also "write" permissions.
-This key can be added using `use_circle_deploy()`.
-This function will add a so called "user key" to the settings of the repo on Circle CI.
+To enable deployment, a "user-key" is needed.
+This key type also has "write" permissions and can be added using `use_circle_deploy()`.
+`use_circle_deploy()` will add a so called "user key" to the settings of the repo on Circle CI.
 The private key will be added to your Github profile under the "SSH and GPG keys" section with the title pointing to the respective repo.
 See also [the section about deployment](circle/deployment.html).
-
-### SSH keys
-
-Circle also makes a distinction between "checkout keys" and "SSH keys", both in the API endpoints and in the Circle CI repo settings (menu SSH permissions).
-Since "checkout keys" are also SSH keys, this creates confusion.
-
-So far I could not figure out what the subtle differences here are and what "normal SSH keys" should be used for.
-If you care about build deployments, you are all set when using "checkout keys".
 
 ---
 
