+++ .github/CONTRIBUTING.md
@@ -1,45 +1,46 @@
-# CONTRIBUTING #
+# CONTRIBUTING
 
 ### Fixing typos
 
 Small typos or grammatical errors in documentation may be edited directly using
 the GitHub web interface, so long as the changes are made in the _source_ file.
 
-*  YES: you edit a roxygen comment in a `.R` file below `R/`.
-*  NO: you edit an `.Rd` file below `man/`.
+- YES: you edit a roxygen comment in a `.R` file below `R/`.
+- NO: you edit an `.Rd` file below `man/`.
 
 ### Prerequisites
 
 Before you make a substantial pull request, you should always file an issue and
 make sure someone from the team agrees that it’s a problem. If you’ve found a
-bug, create an associated issue and illustrate the bug with a minimal 
+bug, create an associated issue and illustrate the bug with a minimal
 [reprex](https://www.tidyverse.org/help/#reprex).
 
 ### Pull request process
 
-*  We recommend that you create a Git branch for each pull request (PR).  
-*  Look at the Travis and AppVeyor build status before and after making changes.
-The `README` should contain badges for any continuous integration services used
-by the package.  
-*  We recommend the tidyverse [style guide](http://style.tidyverse.org).
-You can use the [styler](https://CRAN.R-project.org/package=styler) package to
-apply these styles, but please don't restyle code that has nothing to do with 
-your PR.  
-*  We use [roxygen2](https://cran.r-project.org/package=roxygen2).  
-*  We use [testthat](https://cran.r-project.org/package=testthat). Contributions
-with test cases included are easier to accept.  
-*  For user-facing changes, add a bullet to the top of `NEWS.md` below the
-current development version header describing the changes made followed by your
-GitHub username, and links to relevant issue(s)/PR(s).
+- We recommend that you create a Git branch for each pull request (PR).
+- Look at the Travis and AppVeyor build status before and after making changes.
+  The `README` should contain badges for any continuous integration services used
+  by the package.
+- We recommend the tidyverse [style guide](http://style.tidyverse.org).
+  You can use the [styler](https://CRAN.R-project.org/package=styler) package to
+  apply these styles, but please don't restyle code that has nothing to do with
+  your PR.
+- We use [roxygen2](https://cran.r-project.org/package=roxygen2).
+- We use [testthat](https://cran.r-project.org/package=testthat). Contributions
+  with test cases included are easier to accept.
+- For user-facing changes, add a bullet to the top of `NEWS.md` below the
+  current development version header describing the changes made followed by your
+  GitHub username, and links to relevant issue(s)/PR(s).
 
 ### Testing the package
 
-This package cannot be tested locally by other people than @pat-s since its functionality relies on private API keys.
-These are stored as secure environment variables on the configured CI providers.
-Hence, to test your build, push your changes and have a look at the CI run whether the test are passing.
+To test the package locally, the following conditions need to be met:
 
-During testing, a git repo (`pat-s/travis-testthat`) is initialized and all tests are run on this repo.
-The reasoning for this is that starting new builds and restarting old builds should not happen in a production repository.
+1. A Circle CI API Token stored in an env var named `R_CIRCLE`
+1. An active repository on Circle CI with access rights
+
+Alternatively changes can be pushed to GitHub and being tested by the CI runner.
+The CI runner operates on the `ropenscilabs/circle` repo with access to the respective secrets during the run.
 
 ### Code of Conduct
 
@@ -48,16 +49,17 @@ Please note that the circle project is released with a
 project you agree to abide by its terms.
 
 ### See rOpenSci [contributing guide](https://devguide.ropensci.org/contributingguide.html)
+
 for further details.
 
 ### Discussion forum
 
[…]
