+++ DESCRIPTION
@@ -7,12 +7,12 @@ Authors@R:
            role = c("aut", "cre"),
            email = "patrick.schratz@gmail.com",
            comment = c(ORCID = "0000-0003-0748-6624"))
-Description: Tools for interacting with the 'Circle CI' API.  Besides
-    executing common tasks such as querying build logs and restarting
-    builds, this package also helps setting up permissions to deploy from
-    builds.
+Description: Tools for interacting with the 'Circle CI' API.
+    Besides executing common tasks such as querying build logs and
+    restarting builds, this package also helps setting up permissions to
+    deploy from builds.
 License: GPL-3
-URL: https://ropenscilabs.github.io/circle,
+URL: https://docs.ropensci.org/circle,
     https://github.com/ropenscilabs/circle
 BugReports: https://github.com/ropenscilabs/circle/issues
 Imports:
