### 2020-11-03 f0d9b6d
polish reference and pkgdown structure
