+++ man-roxygen/repo.R
@@ -1,2 +1,2 @@
-#' @param repo `[string]`\cr
+#' @param repo `[character]`\cr
 #'   The repository slug to use. Must follow the "`user/repo`" structure.
