+++ man-roxygen/remote.R
@@ -1,2 +1,2 @@
-#' @param remote `[string]`\cr
-#'   The Github remote which should be used. Defaults to "origin".
+#' @param remote `[character]`\cr
+#'   The Github remote which should be used.
