+++ man-roxygen/user.R
@@ -1,1 +1,2 @@
-#' @param user The username for the repository. By default queried using `get_user()`.
+#' @param user `[character]`\cr
+#'  The username for the repository. By default queried using `get_user()`.
