+++ man-roxygen/vcs.R
@@ -1,2 +1,3 @@
-#' @param vcs_type The version control system to use. Defaults to "gh" (Github).
-#' Other possible values are "bitbucket".
+#' @param vcs_type `[character]`\cr The version control system to use.
+#' Defaults to "gh" (Github).
+#  Other possible values are "bitbucket".
