+++ man-roxygen/api_version.R
@@ -1,2 +1,3 @@
-#' @param api_version A character string specifying the Circle CI API version.
-#'   This usually does not need to be changed by the user.
+#' @param api_version `[character]`\cr
+#' A character string specifying the Circle CI API version.
+#' This usually does not need to be changed by the user.
