+++ R/general.R
@@ -14,11 +14,10 @@ get_circle_user <- function() {
   return(out)
 }
 
-#' @title List repository
-#' @description Retrieve a list of Circle CI repository for the authenticated
+#' @title List Circle CI Projects
+#' @description Retrieve a list of Circle CI repositories for the authenticated
 #'   user.
-#' @param repo The Circle CI project. By default the current directory name will
-#'   be used.
+#' @template repo
 #' @template user
 #' @details Retrieves a very detailed list of repository and repo-related
 #'   information for all Circle CI repository attached to the current user.
@@ -47,16 +46,14 @@ list_projects <- function(repo = github_info()$name,
   return(resp)
 }
 
-#' @title Get build artifacts
+#' @title Get Build Artifacts of a Specific Job
 #' @description Retrieve artifacts from a specific build.
-#' @details Retrieves details about artifacts from a specific build.
-#' @param job_id A Circle CI job id.
+#' @param job_id `[character]`\cr
+#'   A Circle CI job id.
+#' @template repo
 #' @template user
-#' @param repo The Circle CI project. By default the current directory name will
-#'   be used.
 #' @template vcs
 #' @template api_version
-#' @return A list of build artifacts
 #' @examples
 #' \dontrun{
 #' job_id <- get_jobs()[[1]]$id
@@ -85,11 +82,10 @@ get_build_artifacts <- function(job_id = NULL,
   return(resp)
 }
 
-#' @title Trigger build
-#' @description Trigger a new build for a specific repo branch
+#' @title Trigger a New Build on Circle CI
+#' @description Triggers a new build for a specific repo branch.
 #' @details Trigger a new Circle CI build for a specific repo branch.
-#' @param repo The Circle CI project. By default the current directory name will
-#'   be used.
+#' @template repo
 #' @template user
 #' @template vcs
 #' @param branch A character string specifying the repository branch.
@@ -122,10 +118,9 @@ new_build <- function(repo = github_info()$name,
 }
 
 #' @title Enable a repo on Circle CI
-#' @description Follows a repo on Circle CI so that builds can be triggered
+#' @description "Follows" a repo on Circle CI so that builds can be triggered.
 #' @importFrom cli cli_text
-#' @param repo The Circle CI project. By default the current directory name will
-#'   be used.
+#' @template repo
 #' @template user
 #' @template vcs
 #' @template api_version
@@ -154,7 +149,7 @@ enable_repo <- function(repo = github_info()$name,
     sprintf("enable repo %s on Circle CI", repo)
   )
 
-  cli_text("Successfully enabled repo '{user}/{repo}' on Circle CI.")
+  cli::cli_text("Successfully enabled repo '{user}/{repo}' on Circle CI.")
 
   return(invisible(TRUE))
 }
