+++ man/circle-package.Rd
@@ -1,5 +1,5 @@
 % Generated by roxygen2: do not edit by hand
-% Please edit documentation in R/circle-package.R, R/circle.R
+% Please edit documentation in R/circle-package.R
 \docType{package}
 \name{circle-package}
 \alias{circle-package}
@@ -20,12 +20,10 @@ Use of this package requires a Circle API key. Unless a key is already set,
 users will be guided through the creation of a key,
 API keys are disposable, but should still be treated securely.
 
-Use \code{\link[=github_repo]{github_repo()}} to get the name of the current repository
-as determined from the \code{origin} remote.
-The following functions simplify integrating R package testing and deployment
-with GitHub and Circle CI:
+The following functions simplify integrating R package testing and
+deployment with GitHub and Circle CI:
 \itemize{
-\item \code{\link[=circle_enable]{circle_enable()}} enables Circle CI for your repository,
+\item \code{\link[=enable_repo]{enable_repo()}} enables Circle CI for your repository,
 \item \code{\link[=use_circle_deploy]{use_circle_deploy()}} installs a public deploy key on GitHub and the
 corresponding private key on Circle CI to simplify deployments to GitHub
 from Circle CI.
