+++ DESCRIPTION
@@ -1,5 +1,5 @@
 Package: circle
-Title: R Client Package For Circle CI
+Title: R Client Package for Circle CI
 Version: 0.5.0.9002
 Authors@R:
     person(given = "Patrick",
@@ -7,10 +7,10 @@ Authors@R:
            role = c("aut", "cre"),
            email = "patrick.schratz@gmail.com",
            comment = c(ORCID = "0000-0003-0748-6624"))
-Description: Tools for interacting with the 'Circle CI' API.
-    Besides executing common tasks such as querying build logs and
-    restarting builds, this package also helps setting up permissions to
-    deploy from builds.
+Description: Tools for interacting with the 'Circle CI' API.  Besides
+    executing common tasks such as querying build logs and restarting
+    builds, this package also helps setting up permissions to deploy from
+    builds.
 License: GPL-3
 URL: https://docs.ropensci.org/circle/,
     https://github.com/ropenscilabs/circle
