### 2020-11-03 3e95dfc
Bump version to 0.6.0
