### 2020-11-03 #356 @pat-s
@noamross 👋
Happy to announce that the package is ready for review (again). Sorry that it took so long - a move between countries and COVID resulted in little time for FOSS packages, especially for niche ones.

I've completely revamped the test suite:

- now using {vcr}
- For local testing, env vars can be set (user, repo, api token) making it possible for reviewers to test against one of their repos
- Secrets like API keys are hidden in the http mocking approach of {vcr}

In addition, the following enhancements were made:

- Documentation now uses roxygen markdown
- testthat 3.0 ready
- pkgdown reference structure
- updated screenshots
- added parameter types to help pages
- added pre-commit hooks
- added codemeta

Also I've bumped the version to 0.6.0 (maybe this needs to be updated in the first post).
