### 2020-11-05 #356 @noamross
Thanks, @pat-s! I have run local tests and `goodpractice::gp()` checks and this passes with flying colors.  I note that I had to run `use_circle_deploy()` once locally in addition to having my `R_CIRCLE` API key in the repository in order for local test flow to run.   I'm now seeking reviewers.
