### 2020-12-03 #356 @sharlagelfand
- I have not had any working relationship with @pat-s.

---

I am very excited by the potential for this package! I have primarily used Travis and Github Actions in the past but have one project where I need to use Circle CI, and when I set it up wished there was a package to automate things the way that the {travis} and {ghactions} packages do. I believe that with improved documentation, this package will be very helpful for people wanting to automate their Circle CI setup and usage.

I have not checked off the majority of the "Documentation" section, with thoughts described in the sections below, but "Functionality" wise all looks good to me. Beyond what is mentioned below re: examples failing, I did not have trouble using the main functionality (getting jobs, starting builds, etc) and the test suite passed for me (I see above lots of discussion on getting it going - so thank you for that!).

### Overall documentation

Much of the functionality in this package seems solid, but I think it is lacking in the documentation side of things, as Noam mentioned earlier. Neither the README nor the Get Started vignette provide much background into what Circle is or what Continuous Integration is for - you actually have great introductory documentation on this in [`?circle-package`](https://github.com/ropenscilabs/circle/blob/main/R/circle-package.R) so I would love to see that appear elsewhere! As Noam said, multiple points of entry are important, and I don't know if I would have found that page if I was not doing the package review.

On a related note, I would also love to see this page pop up when you do `?circle`, rather than getting the documentation for the [`circle` function](https://ropenscilabs.github.io/circle/reference/circle.html) - would you consider renaming the function and aliasing the documentation in `circle-package` to `circle` as well, since `circle` is explicitly never meant to be called by users?

It would be really great to see a few different vignettes for this package: one for existing users of Circle CI on how to automate their builds, etc, and one for newer users (or existing users who want to set up new repositories) on how to set up the entire thing, *including* how to use {circle} and {tic} in combination. The documentation states that the {tic} package should be used for setting up the YAML file, but does not give any more details or example of how this should be done. This content already has some start in the [Quickstart section of the {tic} README](https://github.com/ropensci/tic#quickstart) but I believe it would be helpful to see it in both places. There is similarly great info on getting started in the [Details section of `use_circle_deploy()`](https://github.com/ropenscilabs/circle/blob/main/R/use-circle-deploy.R#L70) that could be brought more to the forefront.

I also think the [Authentication section in "Get started"](https://ropenscilabs.github.io/circle/articles/circle.html#authentication) would benefit from more details (or even a screen shot, like in the Deployment section below) on how to get the API key. I found it difficult to struggle through the different authentication pieces, and a more detailed walk through would be great!

I found the functions very clear to use, with nicely written code and clear individual documentation! I really like how the arguments show what type they are (e.g. stating when an argument is a character versus a list) - this isn't something I've seen before.

### Examples

Several examples from exported functions failed, listed below.

#### browse_circle_token

`browse_github_token()` errors below because it is part of the {usethis} package, which is not loaded in the example. Also, the examples section (shown below) only uses `browse_github_token()` and `edit_circle_config()`, but not `browse_circle_token()` itself.

``` r
library(circle)

browse_github_token()
# Error in browse_github_token() :
#   could not find function "browse_github_token"

edit_circle_config()
# ● Modify '/Users/sharla/.circleci/cli.yml'
```

#### builds

Both `get_workflows()` and `get_jobs()` error because they are not exported, but are listed in the examples.

``` r
library(circle)

pipelines <- get_pipelines()

workflows <- get_workflows()
# Error in get_workflows() : could not find function "get_workflows"

jobs <- get_jobs()
# Error in get_jobs() : could not find function "get_jobs"
```

#### checkout_key

No `fingerprint` argument is given in the example, so the `delete_checkout_key()` function errors.

``` r
library(circle)

delete_checkout_key()
# Error in delete_checkout_key() :
#  Please provide the fingerprint of the key which should be deleted.
```

#### get_build_artifacts

Errors, same as in the builds documentation, because `get_jobs()` is not exported.

``` r
library(circle)

job_id <- get_jobs()[[1]]$id
# Error in get_jobs() : could not find function "get_jobs"
get_build_artifacts(job_id)
# Error in get_build_artifacts(job_id) : object 'job_id' not found
```

### Function usage

I love how chatty and responsive some of the functions are, like [`enable_repo()`](https://github.com/ropenscilabs/circle/blob/main/R/general.R#L152) and [`use_circle_deploy()`](https://github.com/ropenscilabs/circle/blob/main/R/use-circle-deploy.R#L46)! I find it super helpful when functions tell you what they are doing and guide you to next steps. It would be great to see this extended to other functions like `new_build()`, which has important behaviour it could report back on, instead of returning API information, for example below (note I have redacted a couple IDs myself, not the function doing this!):

``` r
library(circle)

new_build()
# $content
# $content$number
# [1] 53
#
# $content$state
# [1] "pending"
#
# $content$id
# [1] [REDACTED]
#
# $content$created_at
# [1] "2020-12-03T20:27:00.596Z"
#
#
# $path
# [1] "/project/gh/data-mermaid/mermaidr/pipeline"
#
# $response
# Response [https://circleci.com/api/v2/project/gh/data-mermaid/mermaidr/pipeline?circle-token=[REDACTED]]
#   Date: 2020-12-03 20:27
#   Status: 201
#   Content-Type: application/json;charset=utf-8
#   Size: 115 B
#
#
# attr(,"class")
# [1] "circle_api"
```

I also noticed that `get_pipelines()` has a default `repo = NULL`, while (almost?) all of the other functions default to `github_info()$name`. It looks like the function actually [converts the `NULL` to `github_info()$name` immediately](https://github.com/ropenscilabs/circle/blob/main/R/builds.R#L30), so may be worth updating for consistency.

I hope this is helpful - please let me know if you'd like me to clarify any points!
