### 2020-12-08 #356 @mbjoseph
---

Overall this package seems super useful for experienced Circle CI users, and there are a few small things that could be done to broaden the user base and make it easier for folks to get started. I'll echo the suggestions above about expanding the docs to include a bit more background on why this package is useful. Specific suggestions are provided below. At a high level, I like that the package provides good support for common operations with relatively high level abstractions, while also providing users access to the API via the general purpose `circle::circle` function. That's not an easy balance to strike, but I think this package does a good job. Finally, the code itself is clearly written and easy to follow - nice work!

*Specific suggestions*

- The documentation could present use cases for R users (maybe related to R package development, or websites built with R). This could be described in the README at a high level, and then in detail in a vignette. This would help would-be users understand what they get out of the package. Currently, with the way the README is written, it seems like users who don't know what the Circle CI REST API does might not understand whether they would benefit from using the package.

- The package documentation mentions deployments in a number of places. This might be a place where a novice gets tripped up. Is it worth demonstrating the functionality of the package for some deployment (maybe a website or shiny app) in a vignette?

- In [this message](https://github.com/ropenscilabs/circle/blob/3e95dfcc5776dccaeb45ba0047b78d722018444d/R/checkout-key.R#L38-L46) it would be better to mention that the Project Settings > SSH Keys > ... stuff is on Circle CI's website, e.g., "On your repo's Circle CI project page, go to Project Settings > SSH Keys > ...".

- As Sharla mentioned some of the examples are failing because they call functions that aren't exported. It seems like part of the problem is that these examples are wrapped in `dontrun`, so they aren't raising errors upon checking, and another part of the problem is that some functions (e.g., `get_jobs()`) aren't exported. Both parts might point to different solutions, but either way it would be better for a user if the examples complete successfully.

- It would be helpful to document what functions return using `@return`, as this might not be obvious to users. For example, a user might want to `get_build_artifacts`, and the documentation can tell them what to expect as a return value when using that function.
