### 2020-12-12 ed6a68b
update tic templates
