+++ .github/workflows/tic.yml
@@ -1,5 +1,5 @@
 ## tic GitHub Actions template: linux-macos-windows-deploy
-## revision date: 2020-08-06
+## revision date: 2020-12-11
 on:
   workflow_dispatch:
   push:
@@ -28,6 +28,12 @@ jobs:
           - { os: ubuntu-latest, r: "release" }
 
     env:
+      # [Custom env var] Circle CI GITHUB PAT
+      PAT_GITHUB: ${{ secrets.PAT_GITHUB }}
+      # [Custom env var] Circle CI API key
+      CIRCLE_CHECKOUT_KEY: ${{ secrets.CIRCLE_CHECKOUT_KEY }}
+      # [Custom env var] Circle CI API key
+      R_CIRCLE: ${{ secrets.R_CIRCLE }}
       # otherwise remotes::fun() errors cause the build to fail. Example: Unavailability of binaries
       R_REMOTES_NO_ERRORS_FROM_WARNINGS: true
       CRAN: ${{ matrix.config.cran }}
@@ -42,15 +48,9 @@ jobs:
       SDKROOT: /Library/Developer/CommandLineTools/SDKs/MacOSX.sdk
       # use GITHUB_TOKEN from GitHub to workaround rate limits in {remotes}
       GITHUB_PAT: ${{ secrets.GITHUB_TOKEN }}
-      # [Custom env var] Circle CI API key
-      R_CIRCLE: ${{ secrets.R_CIRCLE }}
-      # [Custom env var] Circle CI API key
-      CIRCLE_CHECKOUT_KEY: ${{ secrets.CIRCLE_CHECKOUT_KEY }}
-      # [Custom env var] Circle CI GITHUB PAT
-      PAT_GITHUB: ${{ secrets.PAT_GITHUB }}
 
     steps:
-      - uses: actions/checkout@v2.3.1
+      - uses: actions/checkout@v2.3.4
 
       - uses: r-lib/actions/setup-r@master
         with:
@@ -76,7 +76,7 @@ jobs:
 
       - name: "[Cache] Cache R packages"
         if: runner.os != 'Windows'
-        uses: pat-s/always-upload-cache@v2.1.0
+        uses: pat-s/always-upload-cache@v2.1.3
         with:
           path: ${{ env.R_LIBS_USER }}
           key: ${{ runner.os }}-r-${{ matrix.config.r }}-${{steps.date.outputs.date}}
@@ -88,9 +88,21 @@ jobs:
           Rscript -e "if (!requireNamespace('remotes')) install.packages('remotes', type = 'source')"
           Rscript -e "if (getRversion() < '3.2' && !requireNamespace('curl')) install.packages('curl', type = 'source')"
 
-      - name: "[Stage] [Linux] Install curl"
+      - name: "[Stage] [Linux] Install curl and libgit2"
         if: runner.os == 'Linux'
-        run: sudo apt install libcurl4-openssl-dev
+        run: sudo apt install libcurl4-openssl-dev libgit2-dev
+
+      - name: "[Stage] [macOS] Install libgit2"
+        if: runner.os == 'macOS'
+        run: brew install libgit2
+
+      - name: "[Stage] [macOS] Install system libs for pkgdown"
+        if: runner.os == 'macOS' && matrix.config.pkgdown != ''
+        run: brew install harfbuzz fribidi
+
+      - name: "[Stage] [Linux] Install system libs for pkgdown"
+        if: runner.os == 'Linux' && matrix.config.pkgdown != ''
+        run: sudo apt install libharfbuzz-dev libfribidi-dev
 
       - name: "[Stage] Install"
         if: matrix.config.os != 'macOS-latest' || matrix.config.r != 'devel'
@@ -112,7 +124,7 @@ jobs:
 
[…]
