### 2020-12-16 ropenscilabs/circle#14 @pat-s
Add an advanced vignette
Explaining advanced usage + interaction with {tic)

Move introductory parts to "Get Started" vignette.
