### 2020-12-16 ropenscilabs/circle#15 @pat-s
Fix examples
Due to being wrapped in `dontrun` and the fact that R CMD Checks ignores them by default, errors are not catched.

Atm some functions are not exported or require suggested packages.
