### 2020-12-16 ropenscilabs/circle#16 @pat-s
Check default argument of `get_pipelines()`
All other `get_*` funs have `github_info()$name` as default while `get_pipelines()` has `repo = NULL`.
