### 2020-12-16 #356 @pat-s
@sharlagelfand Thanks for your review!
Please note upfront that even though some of my comments might eventually sound a bit harsh, they are not meant to be like that but are just straight to the point :)

> Neither the README nor the Get Started vignette provide much background into what Circle is or what Continuous Integration is for

The package is a client package - so I'd argue that users arriving here should know what they want to do with it. As a comparable example: a client package wrapping `git` (e.g. [gert](https://github.com/r-lib/gert)) also does not introduce `git` from scratch.
While I see the point of explaining function usage and returns in great detail, I do not think that this applies in the context of the upstream library/service/provider.
Links and cross-refs can help here but I do not want to re-explain CI.
I partly did this even in {tic} (re-explaining CI) but this package has arguably a much more wider scope than this client package here.

> would you consider renaming the function and aliasing the documentation in circle-package to circle as well, since circle is explicitly never meant to be called by users?

Many packages name their workhorse function like the package name.
Also `<package-name>-package` is a common naming scheme for package "intro pages".
How about adding a top-level cross-ref to `circle` which states something like "for a package overview please have a look at XY"?
Given that this would only be a cosmetic change for users it would require quite some code reshuffling work.

> It would be really great to see a few different vignettes for this package: one for existing users of Circle CI on how to automate their builds, etc, and one for newer users (or existing users who want to set up new repositories) on how to set up the entire thing, including how to use {circle} and {tic} in combination.

I agree that this would be benefitial.
However in practice the number of R users on Circle CI is probably less than 10 ;)
https://github.com/ropenscilabs/circle/issues/14

> The documentation states that the {tic} package should be used for setting up the YAML file, but does not give any more details or example of how this should be done. This content already has some start in the Quickstart section of the {tic} README but I believe it would be helpful to see it in both places.

Yes, documentation for the {tic} integration could be improved.
However, {tic} is choice (and probably a biased one) so I want to be stay objective as possible and not bundle the packages too much.

> There is similarly great info on getting started in the Details section of use_circle_deploy() that could be brought more to the forefront.

I will have a look how to bridge the gap between not repeating the help pages and not being to sparse in the vignettes.
I am not a friend of literally repeated text in help pages/vignettes but believe that vignettes should explain the bigger picture and help pages should provide more technical detail of a function.

> I also think the Authentication section in "Get started" would benefit from more details (or even a screen shot, like in the Deployment section below) on how to get the API key.

While I see that this might be helpful in rare instances, I am not sure if I want to go down the "screenshot-explain-everthing" route. I do not see that much value in screenshot-explaining how to copy/create an API-key/Access token from a service.
Screenshots inherit maintenance work since they get outdate quickly if the UI gets updated, hence I try to avoid them unless they are really necessary.

> I really like how the arguments show what type they are (e.g. stating when an argument is a character versus a list) - this isn't something I've seen before.

Thanks. Some packages have this but it always comes down to the maintainers. Definitive not a standard in R.

> Several examples from exported functions failed, listed below.

Thanks, I'll have a look.
https://github.com/ropenscilabs/circle/issues/15

> No fingerprint argument is given in the example, so the delete_checkout_key() function errors.

This is on purpose and that is why it is wrapped in `dontrun`. Most examples cannot run on CRAN due to API calls.
This particular function has no default argument because it cannot delete a key without a reference.

> this extended to other functions like new_build(), which has important behaviour it could report back on,

Which important behaviour?

> I also noticed that get_pipelines() has a default repo = NULL, while (almost?) all of the other functions default to github_info()$name. It looks like the function actually converts the NULL to github_info()$name immediately, so may be worth updating for consistency.

Thanks, might be an oversight.
https://github.com/ropenscilabs/circle/issues/16
