### 2020-12-16 a13c835
- wrap examples in donttest{} and export more get_* functions
