+++ R/circle-package.R
@@ -21,7 +21,7 @@
 #'   corresponding private key on Circle CI to simplify deployments to GitHub
 #'   from Circle CI.
 #' @examples
-#' \dontrun{
+#' \donttest{
 #' # check to see if you've authenticated correctly
 #' get_circle_user()
 #' }
