+++ R/auth.R
@@ -30,8 +30,8 @@
 #'      ```
 #' @export
 #' @examples
-#' \dontrun{
-#' browse_github_token()
+#' \donttest{
+#' browse_circle_token()
 #'
 #' edit_circle_config()
 #' }
