+++ R/auth.R
@@ -30,7 +30,7 @@
 #'      ```
 #' @export
 #' @examples
-#' \donttest{
+#' \dontrun{
 #' browse_circle_token()
 #'
 #' edit_circle_config()
