+++ man/use_circle_deploy.Rd
@@ -33,7 +33,7 @@ access token needs to be set (via env var \code{GITHUB_TOKEN}).
 If none is set, this function will prompt you to create one.
 }
 \examples{
-\donttest{
+\dontrun{
 use_circle_deploy()
 }
 }
