### 2020-12-21 #356 @sharlagelfand
Hey @pat-s, thanks for your response and for creating the linked issues!

I don't think that you necessarily need to explain what Continuous Integration is, but it would be helpful for the entry README and Get Started vignette to at least contain the phrase "Continuous Integration" - I like level of detail in {usethis} [`use_travis` ](https://usethis.r-lib.org/reference/ci.html) and [`use_actions`](https://usethis.r-lib.org/reference/github_actions.html), personally.

Re: vignettes and re-explaining things, I'll refer to the rOpenSci Packages book's [section on documentation](https://devguide.ropensci.org/building.html#documentation), namely (paraphrasing is mine):

* Packages should contain a vignette illustrating realistic use cases
* If your package connects to an online service, it should provide enough information for users to understand the nature of the service - any vignette should outline prerequisite knowledge to understand the vignette upfront.
* README, top-level docs, vignettes, etc should have enough information to get a high-level overview and to follow the principle of multiple points of entry (which Noam referenced)

> Yes, documentation for the {tic} integration could be improved.
> However, {tic} is choice (and probably a biased one) so I want to be stay objective as possible and not bundle the packages too much.

Re: integration with {tic} - that's fine with me if you don't want to bundle them. Would you be able to, instead, include or link an example of a YAML template so that users can get set up?

> This is on purpose and that is why it is wrapped in dontrun. Most examples cannot run on CRAN due to API calls.
> This particular function deletes an API key.
> To delete an item, the item needs to be created first/exist already.

My understanding of the example section is that it contains things a user would actually be able to run - it should illustrate actual usage and not only include a function that you expect to fail. Being wrapped in `dontrun` because it will fail on CRAN is one thing - failing locally is another.

If an item needs to first be created, could the example instead be:

* Create the item (and store the ID in an object)
* Delete the item

Would that work? Related, I realized I don't even know what a "checkout key" is in Circle CI - the only thing that comes up on google is the {circle} package itself. Could you link to the relevant documentation?

> Which important behaviour?

Sorry, was too vague here - important behaviour meaning it triggers a build. As a user, I would be more receptive to and better understand a function that returns e.g. "✓ New build started" rather than API call information.

I imagine that you and other folks will also be taking some time off for the holidays - I will be mostly offline until December 29, so will respond to anything new after that 😊 Happy holidays!
