### 2021-01-07 ropenscilabs/circle#18 @pat-s
Expand "Getting Started" vignette
As suggested during [ropensci review](https://github.com/ropensci/software-review/issues/356#issuecomment-741084556).

- [x] Elaborate on the existence of `circle()` and high-level functions
- [x] Think about screenshots
