### 2021-01-07 ropenscilabs/circle#19 @pat-s
Improve the message if GitHub authentication failed
https://github.com/ropenscilabs/circle/blob/3e95dfcc5776dccaeb45ba0047b78d722018444d/R/checkout-key.R#L38-L46

https://github.com/ropensci/software-review/issues/356#issuecomment-741084556
