### 2021-01-07 ropenscilabs/circle#20 @pat-s
Document return values of functions
https://github.com/ropensci/software-review/issues/356#issuecomment-741084556
