### 2021-01-07 df35543
- optimize message for GitHub authentication issues (fixes Improve the message if GitHub authentication failed #19)
