### 2021-01-07 f2d5ecc
- `get_pipelines()`: set default for arg `repo` according to all other funs (fixes Check default argument of `get_pipelines()` #16)
