### 2021-01-07 0d6d337
- document return values, be more chatty for side-effects, always return a `circle_api` object (fixes Document return values of functions #20 )
