### 2021-01-07 83bbcf1
extend README and getting started vignette
