+++ vignettes/circle.Rmd
@@ -1,5 +1,5 @@
 ---
-title: "circle"
+title: "Getting Started"
 output: rmarkdown::html_vignette
 vignette: >
   %\VignetteIndexEntry{circle}
@@ -14,15 +14,21 @@ knitr::opts_chunk$set(
 )
 ```
 
-# Authentication
+# Introduction
 
-To get started with Circle CI on Github, make sure to install the [Circle CI Github App](https://github.com/marketplace/circleci) and configure the "Free" plan for your account.
+To be able to run builds from a repository, it needs to be enabled on Circle CI first.
+This package heavily relies on GitHub for automatic user information extraction and other providers such as GitLab or Bitbucket are not supported.
+In general Circle CI builds can be run from any Git hosting provider.
+
+# Authentication with GitHub
+
+To get started with Circle CI on GitHub, make sure to install the [Circle CI GitHub App](https://GitHub.com/marketplace/circleci) and configure the "Free" plan for your account.
 
 The first time any {circle} function is called, it will check for the existence of a Circle API key.
-This API key is needed to securely talk to the Circle CI API and make requests.
-There are two ways you can store the API key:
+This API key is needed to securely talk to the Circle CI API and make requests on behalf of a user.
+There are two ways the API key can be set:
 
-  - Via environment variable `R_CIRCLE` in `~/.Renviron` (**recommended**)
+  - (**recommended**) Via environment variable `R_CIRCLE` in `~/.Renviron` (e.g. `R_CIRCLE = <API KEY>`)
   - In `~/.circleci/cli.yml` via the following pattern
 
    ```yml
@@ -31,13 +37,64 @@ There are two ways you can store the API key:
    token: <token>
    ```
 
-In addition, {circle} scrapes information about the current repository from Github.
-To be able to do this, a `GITHUB_TOKEN` is needed.
-You can create one using `usethis::browse_github_token()` if you haven't done so yet.
+To create an API key from the R console, `browse_circle_token()` can be used.
+
+{circle} scrapes information about the current repository by making API calls to GitHub.
+To be able to do this, a `GitHub_TOKEN` is needed (similar to the API requests for Circle CI).
+It is good practice to have such a token set for many purposes when using R with GitHub.
+Invoking `usethis::browse_GitHub_token()` is an easy way to create one if none exists yet.
+
+# First steps
+
+By querying information about the user and enabling a project one can check if everything was set up correctly.
+
+`get_circle_user()` makes uses of the Circle CI API key.
+
+
+```r
+circle::get_circle_user()
+
+$content
+$content$name
+[1] "Patrick Schratz"
+
+$content$login
+[1] "pat-s"
+
+$content$id
+[1] "9c373331-d0f7-45e1-afe6-4a5c75e00d10"
+
+
+$path
+[1] "/me"
+
+$response
+Response [https://circleci.com/api/v2/me?circle-token=39d697f345d8d8a92ab07c333405d9b0092d116c]
+  Date: 2021-01-07 20:07
+  Status: 200
+  Content-Type: application/json;charset=utf-8
+  Size: 86 B
+
+
+attr(,"class")
+[1] "circle_user"
+```
+
+`enable_repo()` also uses the GitHub token.
+
+```r
+circe::enable_repo()
+✔ Successfully enabled repo 'ropenscilabs/circle' on Circle CI.
+```
+
+After the repo has been enabled, it should be returned in `circle::list_projects()`.
 
 # Deployment
 
-The easiest way to enable deployment from Circle CI builds to Github repositories is by using `use_circle_deploy()`:
+Deployments refers to the practice to push (modified) files during a build to a repository.
+Configuring build deployments can be a bit tedious with respect to permissions.
[…]
