### 2021-01-07 ropenscilabs/circle#20 @pat-s
Fixed in 0d6d3375065e0da73920dddcbe9d790b14e82d45
