### 2021-02-09 9fe4913
README: explain CI in more detail and add more function reference links
