### 2021-02-18 ropenscilabs/circle#21 @pat-s
Better explain what a "checkout key" is
source: https://github.com/ropensci/software-review/issues/356#issuecomment-749181931

In the function description of https://docs.ropensci.org/circle/reference/checkout_key.html?q=checkout%20key.
