### 2021-02-18 ropenscilabs/circle#22 @pat-s
Double-check chatty output of functions
E.g. functions like `new_build()` could print more descriptive console output rather than just returning the API response.

source: https://github.com/ropensci/software-review/issues/356#issuecomment-749181931
