### 2021-02-18 #356 @pat-s
@sharlagelfand

> Re: integration with {tic} - that's fine with me if you don't want to bundle them. Would you be able to, instead, include or link an example of a YAML template so that users can get set up?

Thanks. As you've already seen there is a new vignette about [tic & circle](https://docs.ropensci.org/circle/articles/tic.html).

> My understanding of the example section is that it contains things a user would actually be able to run - it should illustrate actual usage and not only include a function that you expect to fail. Being wrapped in dontrun because it will fail on CRAN is one thing - failing locally is another.
If an item needs to first be created, could the example instead be:
Create the item (and store the ID in an object)
Delete the item

I see your point here and were arguably a bit lazy with including meaningful examples.

> Would that work? Related, I realized I don't even know what a "checkout key" is in Circle CI - the only thing that comes up on google is the {circle} package itself. Could you link to the relevant documentation?

Interesting. I'll add some more info to the function help page.

> Sorry, was too vague here - important behaviour meaning it triggers a build. As a user, I would be more receptive to and better understand a function that returns e.g. "✓ New build started" rather than API call information.

Ok thanks, I see.
I think the API information is important and contains metadata which users might want to use for certain postprocessing steps.
I think we could combine both - returning the API info as now and add some verbose {cli} outputs.
