### 2021-02-18 b6e6cc6
add more information about "checkout keys" to the function help page and link to vignette
