+++ R/checkout-key.R
@@ -1,14 +1,16 @@
 #' @title Interact with "Checkout Keys" on Circle CI
-#' @description Create, delete, query or check different types of checkout keys
-#'   for a specific Circle CI project.
-#'   Valid values for argument `type` are `"user-key"` or `"checkout-key"`.
+#' @description
+#' Create, delete, query or check different types of checkout keys
+#' for a specific Circle CI project.
+#' Valid values for argument `type` are `"user-key"` or `"deploy-key"`.
 #'
 #' A "Checkout Key" on Circle CI is a specific SSH key which is used to checkout
 #' repositories into a Circle CI build and possible deploy changes to the
 #' repository.
-#' Circle CI subdivides "Checkout Keys" into "user key" and "deploy key".
+#' Circle CI subdivides "Checkout Keys" into "user-key" and "deploy-key".
 #'
-#' Please see "Deployment" section in the "Getting Started" vignette for more information.
+#' Please see "Deployment" section in the "Getting Started" vignette for more
+#' information.
 #' @template repo
 #' @template user
 #' @template vcs
@@ -30,7 +32,7 @@
 #' # deployments
 #' create_checkout_key()
 #'
-#' # A "deploy key" can only be used to checkout code from the repository into
+#' # A "deploy-key" can only be used to checkout code from the repository into
 #' # a Circle CI build
 #' create_checkout_key(type = "deploy-key")
 #' }
