### 2021-02-18 #356 @pat-s
I've wrapped up all review-related changes in [v0.7.0](https://docs.ropensci.org/circle/news/index.html).

Thanks to @sharlagelfand and @mbjoseph again 👏

@noamross Looks like we can proceed here.
