### 2021-02-18 5b23f9f
clean up NEWS
