### 2021-03-21 #356 @pat-s
Hi @noamross

Just checking base for what might be the next steps from your side - the plan is to release to CRAN after this review is finished.
